Split engine operational tests
