+++ crates/truthdb-core/src/engine/tests.rs
@@ -1,97 +1,16 @@
-use std::path::{Path, PathBuf};
+use std::path::PathBuf;
 
 use super::test_support::*;
+
 use super::*;
 
-/// Stage 2 exit criterion: search events and relational records share
-/// one WAL ring; a crash must recover both, each through its own
-/// mechanism, regardless of interleaving.
-#[test]
-fn mixed_search_and_relational_wal_replays_in_order() {
-    let path = unique_temp_path("mixed-wal");
-    let storage = Storage::create(path.clone(), test_storage_options()).expect("storage create");
-    let engine = Engine::new(storage).expect("engine create");
+mod backup_restore;
 
-    engine
-        .execute(
-            r#"create index docs { "mappings": { "properties": { "body": { "type": "text" } } } }"#,
-        )
-        .expect("create index");
-    engine
-        .execute("CREATE TABLE items (id INT NOT NULL PRIMARY KEY, label NVARCHAR(50))")
-        .expect("create table");
-    // Interleave the two subsystems in one ring.
-    for i in 0..10 {
-        engine
-            .execute(&format!(
-                r#"insert document docs {{ "body": "search event {i}" }}"#
-            ))
-            .expect("insert doc");
-        engine
-            .execute(&format!("INSERT INTO items VALUES ({i}, 'row {i}')"))
-            .expect("insert row");
-    }
-    drop(engine); // crash: everything lives in the shared WAL only
+mod coordinator;
 
-    let storage = Storage::open(path.clone()).expect("reopen");
-    let engine = Engine::new(storage).expect("recover both subsystems");
+mod recovery;
 
-    let response = engine
-        .execute(r#"search docs { "query": { "match": { "body": "search" } } }"#)
-        .expect("search");
-    let response: Value = serde_json::from_str(&response).expect("json");
-    assert_eq!(response["hits"]["total"].as_u64(), Some(10));
-
-    let ids = sql_column_i64(&engine, "SELECT id FROM items ORDER BY id", 0);
-    assert_eq!(
-        ids,
-        (0..10).collect::<Vec<_>>(),
-        "all rows recovered in key order"
-    );
-
-    // Both surfaces stay writable after recovery.
-    engine
-        .execute("INSERT INTO items VALUES (10, 'after recovery')")
-        .expect("insert after recovery");
-    let ids = sql_column_i64(&engine, "SELECT id FROM items WHERE id > 8 ORDER BY id", 0);
-    assert_eq!(ids, vec![9, 10]);
-    let _ = std::fs::remove_file(path);
-}
-
-#[test]
-fn engine_replay_ignores_relational_wal_records() {
-    let path = unique_temp_path("rel-coexistence");
-    let mut storage =
-        Storage::create(path.clone(), test_storage_options()).expect("storage create");
-    // Relational records land in the same ring before and between search
-    // events; search replay must skip them.
-    let extent = storage.allocate_extent(false).expect("extent");
-    let engine = Engine::new(storage).expect("engine create");
-    engine
-        .execute(
-            r#"
-                create index notes {
-                  "mappings": { "properties": { "body": { "type": "text" } } }
-                }
-                "#,
-        )
-        .expect("create index");
-    engine
-        .execute(r#"insert document notes { "body": "relational coexistence" }"#)
-        .expect("insert");
-    drop(engine);
-
-    let storage = Storage::open(path.clone()).expect("reopen");
-    let engine = Engine::new(storage).expect("engine replay with rel records");
-    let response = engine
-        .execute(r#"search notes { "query": { "match": { "body": "coexistence" } } }"#)
-        .expect("search after replay");
-    let response: Value = serde_json::from_str(&response).expect("valid json");
-    assert_eq!(response["hits"]["total"].as_u64(), Some(1));
-    let _ = extent;
-
-    let _ = std::fs::remove_file(path);
-}
+mod replication;
 
 #[test]
 fn sql_create_insert_select_survive_restart() {
@@ -282,2739 +201,6 @@ fn sql_identity_assigns_and_survives_restart() {
         .expect("i4");
     let (_, rows) = sql_rows(&engine, "SELECT id FROM t WHERE name = 'e'");
     assert_eq!(rows, vec![vec![Some("5".into())]]);
-    let _ = std::fs::remove_file(path);
-}
-
-#[test]
-fn full_backup_and_offline_restore_round_trips_relational_data() {
-    let src = unique_temp_path("backup-src");
-    let bak = unique_temp_path("backup-bak");
-    let restored = unique_temp_path("backup-restored");
-
-    let engine = new_engine(&src);
-    engine
-        .execute("CREATE TABLE t (id INT NOT NULL PRIMARY KEY, name NVARCHAR(20))")
-        .expect("create t");
-    engine
-        .execute("CREATE INDEX ix_name ON t (name)")
-        .expect("create index");
-    // Enough rows to spread the heap and the secondary index over several
-    // B+tree pages, so the copy is more than a single catalog page.
-    for i in 1..=200 {
-        engine
-            .execute(&format!("INSERT INTO t (id, name) VALUES ({i}, 'row{i}')"))
-            .expect("insert into t");
-    }
-    engine
-        .execute("CREATE TABLE u (k INT NOT NULL PRIMARY KEY, v INT)")
-        .expect("create u");
-    engine
-        .execute("INSERT INTO u (k, v) VALUES (1, 10), (2, 20)")
-        .expect("insert into u");
-
-    let expected_t = sql_rows(&engine, "SELECT id, name FROM t ORDER BY id").1;
-
-    // Online backup while the engine is live.
-    let summary = engine
-        .storage()
-        .backup_full(&bak)
-        .expect("online full backup");
-    assert!(summary.pages_copied > 0, "the backup copied data pages");
-    assert!(
-        summary.backup_end_lsn >= summary.redo_start_lsn,
-        "the log bracket is well-formed"
-    );
-    drop(engine);
-
-    // Offline restore into a fresh file, then open it and compare.
-    Storage::restore_full(&restored, &bak).expect("offline restore");
-    let engine2 = Engine::new(Storage::open(restored.clone()).expect("open restored"))
-        .expect("engine on restored file");
-
-    assert_eq!(
-        sql_rows(&engine2, "SELECT id, name FROM t ORDER BY id").1,
-        expected_t,
-        "table t round-trips row-for-row"
-    );
-    assert_eq!(
-        sql_rows(&engine2, "SELECT k, v FROM u ORDER BY k").1,
-        vec![
-            vec![Some("1".into()), Some("10".into())],
-            vec![Some("2".into()), Some("20".into())],
-        ],
-        "table u round-trips"
-    );
-    // The secondary index round-trips: a seek by name finds the row.
-    assert_eq!(
-        sql_rows(&engine2, "SELECT id FROM t WHERE name = 'row150'").1,
-        vec![vec![Some("150".into())]],
-        "the secondary index is intact after restore"
-    );
-    // The catalog round-trips: further DML on the restored database works.
-    engine2
-        .execute("INSERT INTO u (k, v) VALUES (3, 30)")
-        .expect("insert into restored u");
-    assert_eq!(
-        sql_rows(&engine2, "SELECT v FROM u WHERE k = 3").1,
-        vec![vec![Some("30".into())]]
-    );
-    drop(engine2);
-
-    let _ = std::fs::remove_file(src);
-    let _ = std::fs::remove_file(bak);
-    let _ = std::fs::remove_file(restored);
-}
-
-#[test]
-fn backup_database_statement_backs_up_and_restores() {
-    let src = unique_temp_path("backup-stmt-src");
-    let bak = unique_temp_path("backup-stmt-bak");
-    let restored = unique_temp_path("backup-stmt-restored");
-
-    let engine = new_engine(&src);
-    engine
-        .execute("CREATE TABLE t (id INT NOT NULL PRIMARY KEY, name NVARCHAR(20))")
-        .expect("create");
-    for i in 1..=50 {
-        engine
-            .execute(&format!("INSERT INTO t (id, name) VALUES ({i}, 'r{i}')"))
-            .expect("insert");
-    }
-    let expected = sql_rows(&engine, "SELECT id, name FROM t ORDER BY id").1;
-
-    // The T-SQL BACKUP statement drives the online backup.
-    let path_lit = bak.to_str().unwrap().replace('\'', "''");
-    let env = sql(
-        &engine,
-        &format!("BACKUP DATABASE truthdb TO DISK = '{path_lit}' WITH CHECKSUM, COPY_ONLY"),
-    );
-    assert!(env["error"].is_null(), "BACKUP DATABASE failed: {env}");
-    drop(engine);
-
-    Storage::restore_full(&restored, &bak).expect("restore");
-    let engine2 =
-        Engine::new(Storage::open(restored.clone()).expect("open restored")).expect("engine");
-    assert_eq!(
-        sql_rows(&engine2, "SELECT id, name FROM t ORDER BY id").1,
-        expected,
-        "the BACKUP-statement backup restores row-for-row"
-    );
-    drop(engine2);
-
-    let _ = std::fs::remove_file(src);
-    let _ = std::fs::remove_file(bak);
-    let _ = std::fs::remove_file(restored);
-}
-
-#[test]
-fn backup_database_is_rejected_inside_a_transaction() {
-    let path = unique_temp_path("backup-in-txn");
-    let engine = new_engine(&path);
-    // BACKUP manages its own per-chunk locking and cannot run inside an
-    // explicit transaction (3021).
-    assert_eq!(
-        sql_error_number(
-            &engine,
-            "BEGIN TRANSACTION; BACKUP DATABASE d TO DISK = '/tmp/truthdb-never.bak'"
-        ),
-        3021
-    );
-    // A side-effecting BACKUP is illegal inside a function body (156);
-    // otherwise a per-row SELECT would run a backup per row.
-    assert_eq!(
-        sql_error_number(
-            &engine,
-            "CREATE FUNCTION dbo.f() RETURNS INT AS BEGIN \
-                 BACKUP DATABASE d TO DISK = '/tmp/truthdb-never.bak'; RETURN 1 END"
-        ),
-        156
-    );
-    drop(engine);
-    let _ = std::fs::remove_file(path);
-}
-
-#[test]
-fn recovery_model_sets_persists_and_reports() {
-    let path = unique_temp_path("recovery-model");
-    let engine = new_engine(&path);
-    let model = |e: &Engine| sql_rows(e, "SELECT recovery_model_desc FROM sys.databases").1;
-
-    // SIMPLE is the default.
-    assert_eq!(model(&engine), vec![vec![Some("SIMPLE".into())]]);
-
-    engine
-        .execute("ALTER DATABASE CURRENT SET RECOVERY FULL")
-        .expect("set full");
-    assert_eq!(model(&engine), vec![vec![Some("FULL".into())]]);
-
-    // An unrelated option in the same statement family leaves it untouched.
-    engine
-        .execute("ALTER DATABASE CURRENT SET READ_COMMITTED_SNAPSHOT ON")
-        .expect("rcsi on");
-    assert_eq!(model(&engine), vec![vec![Some("FULL".into())]]);
-    drop(engine);
-
-    // FULL persists across a reopen (the set is itself durable).
-    let engine2 = Engine::new(Storage::open(path.clone()).expect("reopen")).expect("engine");
-    assert_eq!(model(&engine2), vec![vec![Some("FULL".into())]]);
-    assert_eq!(
-        sql_rows(
-            &engine2,
-            "SELECT is_read_committed_snapshot_on FROM sys.databases"
-        )
-        .1,
-        vec![vec![Some("1".into())]],
-        "RCSI survived alongside the recovery model"
-    );
-
-    engine2
-        .execute("ALTER DATABASE CURRENT SET RECOVERY SIMPLE")
-        .expect("set simple");
-    assert_eq!(model(&engine2), vec![vec![Some("SIMPLE".into())]]);
-    drop(engine2);
-    let _ = std::fs::remove_file(path);
-}
-
-#[test]
-fn backup_log_ships_the_log_advances_the_marker_and_chains() {
-    use crate::backup::{BackupReader, BlockType};
-    let src = unique_temp_path("backuplog-src");
-    let trn1 = unique_temp_path("backuplog-1");
-    let trn2 = unique_temp_path("backuplog-2");
-    let engine = new_engine(&src);
-    engine
-        .execute("ALTER DATABASE CURRENT SET RECOVERY FULL")
-        .expect("full");
-    // Enabling FULL starts the log chain at the current tail and pins it.
-    let chain_start = engine.storage().last_log_backup_lsn();
-    assert_eq!(engine.storage().log_backup_hold(), Some(chain_start));
-    engine
-        .execute("CREATE TABLE t (id INT NOT NULL PRIMARY KEY, v INT)")
-        .expect("create");
-    for i in 1..=20 {
-        engine
-            .execute(&format!("INSERT INTO t VALUES ({i}, {i})"))
-            .expect("insert");
-    }
-
-    let read_log_archive = |path: &std::path::Path| -> (crate::backup::BackupHeader, u64) {
-        let (mut r, header) =
-            BackupReader::new(std::io::BufReader::new(std::fs::File::open(path).unwrap())).unwrap();
-        let mut log_bytes = 0u64;
-        while let Some((bt, payload)) = r.next_block().unwrap() {
-            if bt == BlockType::LogChunk {
-                log_bytes += payload.len().saturating_sub(8) as u64; // minus the start-LSN prefix
-            }
-        }
-        (header, log_bytes)
-    };
-
-    // First BACKUP LOG ships [chain_start, tail) and advances the marker.
-    let lit1 = trn1.to_str().unwrap().replace('\'', "''");
-    assert!(
-        sql(&engine, &format!("BACKUP LOG truthdb TO DISK = '{lit1}'"))["error"].is_null(),
-        "BACKUP LOG succeeded"
-    );
-    let marker1 = engine.storage().last_log_backup_lsn();
-    assert!(marker1 > chain_start, "the marker advanced");
-    assert_eq!(
-        engine.storage().log_backup_hold(),
-        Some(marker1),
-        "the hold moved to the new marker"
-    );
-    let (header1, bytes1) = read_log_archive(&trn1);
-    assert!(header1.flags.log_backup, "flagged as a log-only archive");
-    assert_eq!(header1.redo_start_lsn, chain_start);
-    assert_eq!(
-        chain_start + bytes1,
-        marker1,
-        "the shipped range ends at the marker"
-    );
-
-    // A second BACKUP LOG chains contiguously from the first.
-    engine
-        .execute("INSERT INTO t VALUES (99, 99)")
-        .expect("more");
-    let lit2 = trn2.to_str().unwrap().replace('\'', "''");
-    assert!(sql(&engine, &format!("BACKUP LOG truthdb TO DISK = '{lit2}'"))["error"].is_null());
-    let (header2, _) = read_log_archive(&trn2);
-    assert_eq!(
-        header2.redo_start_lsn, marker1,
-        "the second archive starts where the first ended"
-    );
-    drop(engine);
-    for p in [src, trn1, trn2] {
-        let _ = std::fs::remove_file(p);
-    }
-}
-
-#[test]
-fn restoring_a_full_recovery_database_opens_and_checkpoints_cleanly() {
-    let src = unique_temp_path("restore-full-src");
-    let bak = unique_temp_path("restore-full-bak");
-    let restored = unique_temp_path("restore-full-restored");
-    let trn = unique_temp_path("restore-full-trn");
-    let engine = new_engine(&src);
-    engine
-        .execute("ALTER DATABASE CURRENT SET RECOVERY FULL")
-        .expect("full");
-    engine
-        .execute("CREATE TABLE t (id INT NOT NULL PRIMARY KEY, v INT)")
-        .expect("create");
-    for i in 1..=30 {
-        engine
-            .execute(&format!("INSERT INTO t VALUES ({i}, {i})"))
-            .expect("insert");
-    }
-    let expected = sql_rows(&engine, "SELECT id, v FROM t ORDER BY id").1;
-    engine
-        .storage()
-        .backup_full(&bak)
-        .expect("full backup of a FULL-model db");
-    drop(engine);
-
-    Storage::restore_full(&restored, &bak).expect("restore");
-    let engine2 = Engine::new(Storage::open(restored.clone()).expect("open")).expect("engine");
-    assert_eq!(
-        sql_rows(&engine2, "SELECT id, v FROM t ORDER BY id").1,
-        expected
-    );
-    // The restored DB is FULL and its log-backup floor is seeded at the
-    // restore point (backup_end), so the on-open hold sits at/above wal_head.
-    assert_eq!(
-        sql_rows(&engine2, "SELECT recovery_model_desc FROM sys.databases").1,
-        vec![vec![Some("FULL".into())]]
-    );
-    assert!(
-        engine2.storage().log_backup_hold().is_some(),
-        "FULL-model hold re-registered on the restored db"
-    );
-    // A checkpoint would panic (set_head with a floor below the head) if the
-    // marker had been left at 0 — the Fix-3 regression guard.
-    engine2
-        .storage()
-        .write_checkpoint(b"cp", 1, 2, 1)
-        .expect("checkpoint on the restored db");
-    // And BACKUP LOG works on the restored (fresh) log chain.
-    let lit = trn.to_str().unwrap().replace('\'', "''");
-    assert!(sql(&engine2, &format!("BACKUP LOG truthdb TO DISK = '{lit}'"))["error"].is_null());
-    drop(engine2);
-    for p in [src, bak, restored, trn] {
-        let _ = std::fs::remove_file(p);
-    }
-}
-
-#[test]
-fn restore_full_plus_log_chain_recovers_post_backup_changes() {
-    let src = unique_temp_path("restlog-src");
-    let bak = unique_temp_path("restlog-bak");
-    let trn1 = unique_temp_path("restlog-1");
-    let trn2 = unique_temp_path("restlog-2");
-    let restored = unique_temp_path("restlog-restored");
-    let restored_full_only = unique_temp_path("restlog-fullonly");
-    let restored_gap = unique_temp_path("restlog-gap");
-
-    let engine = new_engine(&src);
-    engine
-        .execute("ALTER DATABASE CURRENT SET RECOVERY FULL")
-        .expect("full");
-    engine
-        .execute("CREATE TABLE t (id INT NOT NULL PRIMARY KEY, v INT)")
-        .expect("create");
-    for i in 1..=20 {
-        engine
-            .execute(&format!("INSERT INTO t VALUES ({i}, {i})"))
-            .expect("insert");
-    }
-    // Full backup captures rows 1..=20.
-    engine.storage().backup_full(&bak).expect("full backup");
-    // Changes AFTER the full backup, then a first log backup.
-    for i in 21..=40 {
-        engine
-            .execute(&format!("INSERT INTO t VALUES ({i}, {i})"))
-            .expect("insert");
-    }
-    let lit1 = trn1.to_str().unwrap().replace('\'', "''");
-    assert!(sql(&engine, &format!("BACKUP LOG truthdb TO DISK = '{lit1}'"))["error"].is_null());
-    // More changes, then a second (chained) log backup.
-    engine
-        .execute("UPDATE t SET v = v + 100 WHERE id <= 5")
-        .expect("update");
-    for i in 41..=50 {
-        engine
-            .execute(&format!("INSERT INTO t VALUES ({i}, {i})"))
-            .expect("insert");
-    }
-    let lit2 = trn2.to_str().unwrap().replace('\'', "''");
-    assert!(sql(&engine, &format!("BACKUP LOG truthdb TO DISK = '{lit2}'"))["error"].is_null());
-    let expected = sql_rows(&engine, "SELECT id, v FROM t ORDER BY id").1;
-    drop(engine);
-
-    // Full + the whole log chain recovers EVERY committed change.
-    Storage::restore_full_with_logs(&restored, &bak, &[trn1.clone(), trn2.clone()], None)
-        .expect("restore full + log chain");
-    let engine2 = Engine::new(Storage::open(restored.clone()).expect("open")).expect("engine");
-    assert_eq!(
-        sql_rows(&engine2, "SELECT id, v FROM t ORDER BY id").1,
-        expected,
-        "restore + log chain recovers all post-full-backup changes"
-    );
-
-    // The full backup alone recovers only the 20 rows at its point.
-    Storage::restore_full(&restored_full_only, &bak).expect("restore full only");
-    let engine3 =
-        Engine::new(Storage::open(restored_full_only.clone()).expect("open")).expect("engine");
-    assert_eq!(
-        sql_rows(&engine3, "SELECT COUNT(*) FROM t").1,
-        vec![vec![Some("20".into())]],
-        "the full backup alone is the point-in-time it was taken"
-    );
-
-    // A gap in the chain (apply only the second log) is rejected (4305), and
-    // the partial destination is removed so a retry can reuse the path.
-    assert!(
-        Storage::restore_full_with_logs(&restored_gap, &bak, &[trn2.clone()], None).is_err(),
-        "a log-chain gap is rejected"
-    );
-    assert!(
-        !restored_gap.exists(),
-        "the partial destination is cleaned up on error"
-    );
-    Storage::restore_full_with_logs(&restored_gap, &bak, &[trn1.clone(), trn2.clone()], None)
-        .expect("retry with the full chain to the same path succeeds after cleanup");
-
-    drop(engine2);
-    drop(engine3);
-    for p in [
-        src,
-        bak,
-        trn1,
-        trn2,
-        restored,
-        restored_full_only,
-        restored_gap,
-    ] {
-        let _ = std::fs::remove_file(p);
-    }
-}
-
-// Stage 18 slice 2: a standby seeded from a full backup, fed the primary's
-// raw shipped WAL ring bytes (`read_wal_range`), recovers to a state that
-// matches the primary — the physical-replication apply path, offline.
-#[test]
-fn standby_applies_shipped_wal_ranges_and_matches_the_primary() {
-    let src = unique_temp_path("repl-src");
-    let bak = unique_temp_path("repl-bak");
-    let standby = unique_temp_path("repl-standby");
-    let standby_idem = unique_temp_path("repl-standby-idem");
-
-    let engine = new_engine(&src);
-    engine
-        .execute("CREATE TABLE t (id INT NOT NULL PRIMARY KEY, v INT)")
-        .expect("create");
-    for i in 1..=20 {
-        engine
-            .execute(&format!("INSERT INTO t VALUES ({i}, {i})"))
-            .expect("insert");
-    }
-    // Full backup captures rows 1..=20 up to `backup_end`.
-    let summary = engine.storage().backup_full(&bak).expect("full backup");
-    let backup_end = summary.backup_end_lsn;
-    // Committed changes AFTER the backup — the log a standby must catch up on.
-    for i in 21..=40 {
-        engine
-            .execute(&format!("INSERT INTO t VALUES ({i}, {i})"))
-            .expect("insert");
-    }
-    engine
-        .execute("UPDATE t SET v = v + 100 WHERE id <= 5")
-        .expect("update");
-    // Ship the primary's raw ring bytes `[backup_end, durable tail)`.
-    let flushed = engine.storage().wal_flushed_lsn();
-    assert!(flushed > backup_end, "there is post-backup log to ship");
-    let delta = engine
-        .storage()
-        .read_wal_range(backup_end, flushed)
-        .expect("ship the WAL range");
-    let expected = sql_rows(&engine, "SELECT id, v FROM t ORDER BY id").1;
-    drop(engine);
-
-    // Standby = the full backup + the shipped raw WAL range, recovered.
-    Storage::restore_full_with_wal_ranges(&standby, &bak, &[(backup_end, delta.clone())])
-        .expect("apply the shipped range to the standby");
-    let s = Engine::new(Storage::open(standby.clone()).expect("open")).expect("engine");
-    assert_eq!(
-        sql_rows(&s, "SELECT id, v FROM t ORDER BY id").1,
-        expected,
-        "the standby matches the primary after applying the shipped WAL"
-    );
-    // The replication restartpoint persisted = the end of the applied range.
-    assert_eq!(
-        s.storage().applied_lsn(),
-        backup_end + delta.len() as u64,
-        "applied_lsn is the end of the applied range and survives the reopen"
-    );
-
-    // Idempotent: applying the SAME range twice yields the identical state
-    // (seed overwrites identical bytes; redo is page-LSN-gated).
-    Storage::restore_full_with_wal_ranges(
-        &standby_idem,
-        &bak,
-        &[(backup_end, delta.clone()), (backup_end, delta.clone())],
-    )
-    .expect("re-applying the same range is accepted");
-    let s2 = Engine::new(Storage::open(standby_idem.clone()).expect("open")).expect("engine");
-    assert_eq!(
-        sql_rows(&s2, "SELECT id, v FROM t ORDER BY id").1,
-        expected,
-        "re-applying the same range is idempotent"
-    );
-
-    drop(s);
-    drop(s2);
-    for p in [src, bak, standby, standby_idem] {
-        let _ = std::fs::remove_file(p);
-    }
-}
-
-// Stage 18 slice 4d: an OPEN standby applies a shipped WAL range LIVE (no
-// reopen) and its state matches the primary; the apply is idempotent and
-// survives a standby restart.
-#[test]
-fn a_standby_applies_a_live_wal_stream_and_matches_the_primary() {
-    let primary_path = unique_temp_path("live-primary");
-    let bak = unique_temp_path("live-bak");
-    let standby_path = unique_temp_path("live-standby");
-
-    // Primary: rows 1..=10, a backup, then post-backup changes to ship.
-    let primary = new_engine(&primary_path);
-    primary
-        .execute("CREATE TABLE t (id INT NOT NULL PRIMARY KEY, v INT)")
-        .expect("create");
-    for i in 1..=10 {
-        primary
-            .execute(&format!("INSERT INTO t VALUES ({i}, {i})"))
-            .expect("insert");
-    }
-    let summary = primary.storage().backup_full(&bak).expect("backup");
-    let backup_end = summary.backup_end_lsn;
-    for i in 11..=25 {
-        primary
-            .execute(&format!("INSERT INTO t VALUES ({i}, {i})"))
-            .expect("insert");
-    }
-    primary
-        .execute("UPDATE t SET v = v + 100 WHERE id <= 5")
-        .expect("update");
-    let flushed = primary.storage().wal_flushed_lsn();
-    let delta = primary
-        .storage()
-        .read_wal_range(backup_end, flushed)
-        .expect("ship the delta");
-    let expected = sql_rows(&primary, "SELECT id, v FROM t ORDER BY id").1;
-
-    // Standby: restore the backup as an OPEN engine (only the 10 backed-up
-    // rows), then apply the shipped delta LIVE — no reopen.
-    Storage::restore_full(&standby_path, &bak).expect("restore");
-    let standby = Engine::new(Storage::open(standby_path.clone()).expect("open")).expect("eng");
-    assert_eq!(
-        sql_rows(&standby, "SELECT COUNT(*) FROM t").1,
-        vec![vec![Some("10".into())]],
-        "the standby starts at the backup point"
-    );
-    standby
-        .storage()
-        .apply_wal_stream(backup_end, &delta)
-        .expect("live apply");
-    assert_eq!(
-        sql_rows(&standby, "SELECT id, v FROM t ORDER BY id").1,
-        expected,
-        "the standby matches the primary after applying the live stream (no reopen)"
-    );
-
-    // Idempotent: re-applying the same range changes nothing.
-    standby
-        .storage()
-        .apply_wal_stream(backup_end, &delta)
-        .expect("re-apply");
-    assert_eq!(
-        sql_rows(&standby, "SELECT id, v FROM t ORDER BY id").1,
-        expected,
-        "re-applying the same range is idempotent"
-    );
-
-    // Durable: the applied state survives a standby restart.
-    drop(standby);
-    let standby2 = Engine::new(Storage::open(standby_path.clone()).expect("reopen")).expect("eng");
-    assert_eq!(
-        sql_rows(&standby2, "SELECT id, v FROM t ORDER BY id").1,
-        expected,
-        "the live-applied state persists across a standby restart"
-    );
-
-    // A backup taken ON the standby captures the FULL applied state (the
-    // in-memory WAL tail is resynced, so the backup is not truncated to the
-    // pre-apply point).
-    let standby_bak = unique_temp_path("live-standby-bak");
-    let restored = unique_temp_path("live-standby-restored");
-    standby2
-        .storage()
-        .backup_full(&standby_bak)
-        .expect("backup on the standby");
-    Storage::restore_full(&restored, &standby_bak).expect("restore the standby backup");
-    let r = Engine::new(Storage::open(restored.clone()).expect("open")).expect("eng");
-    assert_eq!(
-        sql_rows(&r, "SELECT id, v FROM t ORDER BY id").1,
-        expected,
-        "a backup of the standby restores the full applied state"
-    );
-
-    // A checkpoint on a standby is refused (it reclaims ring space only at
-    // promotion, to keep the in-flight undo log).
-    assert!(
-        standby2.checkpoint().is_err(),
-        "checkpoint is refused on a standby"
-    );
-
-    // A standby is read-only: a local client write is rejected (it would
-    // append to the replica's own WAL and diverge it from the primary).
-    let write = sql(&standby2, "INSERT INTO t VALUES (999, 999)");
-    assert!(
-        !write["error"].is_null(),
-        "a local write on a standby is rejected: {write}"
-    );
-    assert_eq!(
-        sql_rows(&standby2, "SELECT COUNT(*) FROM t").1,
-        vec![vec![Some("25".into())]],
-        "the rejected write left the standby unchanged"
-    );
-
-    drop(primary);
-    drop(standby2);
-    drop(r);
-    for p in [primary_path, bak, standby_path, standby_bak, restored] {
-        let _ = std::fs::remove_file(p);
-    }
-}
-
-// Stage 18 slice 4c: the full transport, end to end over a real TCP+TLS
-// socket — listener + handshake + per-standby sender on the primary,
-// reconnecting receiver on the standby. A backup-seeded standby catches up,
-// follows live commits (woken by the flushed watch, no polling), its
-// FlushAcks advance the primary's slot, and a receiver restart resumes from
-// the standby's persisted watermark.
-#[tokio::test(flavor = "multi_thread", worker_threads = 4)]
-async fn the_replication_transport_streams_live_writes_to_a_standby() {
-    use crate::repl::listener::{PrimaryReplContext, run_repl_listener};
-    use crate::repl::receiver::{ReceiverConfig, run_standby_receiver};
-    use crate::repl::tls::server_config_from_pem;
-    use std::sync::Arc;
-    use std::time::Duration;
-    use tokio::sync::watch;
-    use tokio_rustls::TlsAcceptor;
-
-    const SECRET: &[u8] = b"transport-secret";
-    const UUID: [u8; 16] = [7u8; 16];
-
-    let primary_path = unique_temp_path("xport-primary");
-    let bak = unique_temp_path("xport-bak");
-    let standby_path = unique_temp_path("xport-standby");
-
-    // Primary: rows 1..=10, a backup to seed the standby, then post-backup
-    // rows the transport must catch the standby up on.
-    let primary = new_engine(&primary_path);
-    primary
-        .execute("CREATE TABLE t (id INT NOT NULL PRIMARY KEY, v INT)")
-        .expect("create");
-    for i in 1..=10 {
-        primary
-            .execute(&format!("INSERT INTO t VALUES ({i}, {i})"))
-            .expect("insert");
-    }
-    primary.storage().backup_full(&bak).expect("backup");
-    // A catch-up backlog spanning MANY sender chunks (the context below
-    // sets 512-byte chunks), so entry-boundary chunking is exercised end
-    // to end — a mid-entry cut would fail the apply's coverage check.
-    for i in 11..=120 {
-        primary
-            .execute(&format!("INSERT INTO t VALUES ({i}, {i})"))
-            .expect("insert");
-    }
-
-    // The primary's replication listener on an ephemeral port.
-    let (cert_pem, key_pem) = {
-        let c = rcgen::generate_simple_self_signed(vec!["localhost".to_string()]).unwrap();
-        (c.cert.pem(), c.key_pair.serialize_pem())
-    };
-    let acceptor =
-        TlsAcceptor::from(server_config_from_pem(cert_pem.as_bytes(), key_pem.as_bytes()).unwrap());
-    let listener = tokio::net::TcpListener::bind("127.0.0.1:0").await.unwrap();
-    let addr = listener.local_addr().unwrap();
-    let (shutdown_tx, shutdown_rx) = watch::channel(false);
-    let ctx = PrimaryReplContext {
-        shared_secret: Arc::new(SECRET.to_vec()),
-        cluster_uuid: UUID,
-        storage: primary.storage_arc(),
-        heartbeat: Duration::from_millis(200),
-        stall_timeout: Duration::from_secs(30),
-        chunk_bytes: 512,
-    };
-    let listener_task = tokio::spawn(run_repl_listener(
-        listener,
-        acceptor,
-        ctx,
-        shutdown_rx.clone(),
-    ));
-
-    // Standby: seeded with a --standby restore (stamped redo-only +
-    // read-only before its first open), opened live, then its receiver
-    // dials the primary.
-    Storage::restore_full_standby(&standby_path, &bak, &[]).expect("restore");
-    let standby = Engine::new(Storage::open(standby_path.clone()).expect("open")).expect("eng");
-    assert!(
-        standby.storage().is_standby(),
-        "a --standby restore stamps the standby mode before the first open"
-    );
-    let receiver_cfg = ReceiverConfig {
-        primary_addr: addr.to_string(),
-        server_name: "localhost".to_string(),
-        tls_ca_pem: cert_pem.as_bytes().to_vec(),
-        shared_secret: SECRET.to_vec(),
-        cluster_uuid: UUID,
-        node_id: 7,
-        reconnect_delay: Duration::from_millis(100),
-        stall_timeout: Duration::from_secs(30),
-    };
-    let (rx_shutdown_tx, rx_shutdown_rx) = watch::channel(false);
-    let receiver_task = tokio::spawn(run_standby_receiver(
-        receiver_cfg.clone(),
-        standby.storage_arc(),
-        rx_shutdown_rx,
-    ));
-
-    async fn wait_until(what: &str, mut cond: impl FnMut() -> bool) {
-        let deadline = std::time::Instant::now() + Duration::from_secs(20);
-        while !cond() {
-            assert!(
-                std::time::Instant::now() < deadline,
-                "timed out waiting for {what}"
-            );
-            tokio::time::sleep(Duration::from_millis(25)).await;
-        }
-    }
-
-    // Catch-up: the standby reaches the primary's durable watermark.
-    let target = primary.storage().wal_flushed_lsn();
-    wait_until("catch-up", || standby.storage().wal_tail() >= target).await;
-    let expected = sql_rows(&primary, "SELECT id, v FROM t ORDER BY id").1;
-    assert_eq!(
-        sql_rows(&standby, "SELECT id, v FROM t ORDER BY id").1,
-        expected,
-        "the standby matches the primary after catch-up"
-    );
-
-    // Live follow: new commits arrive without any reconnect.
-    for i in 21..=30 {
-        primary
-            .execute(&format!("INSERT INTO t VALUES ({i}, {i})"))
-            .expect("insert");
-    }
-    primary
-        .execute("UPDATE t SET v = v + 100 WHERE id <= 5")
-        .expect("update");
-    let target = primary.storage().wal_flushed_lsn();
-    wait_until("live follow", || standby.storage().wal_tail() >= target).await;
-    let expected = sql_rows(&primary, "SELECT id, v FROM t ORDER BY id").1;
-    assert_eq!(
-        sql_rows(&standby, "SELECT id, v FROM t ORDER BY id").1,
-        expected,
-        "the standby follows live commits"
-    );
-
-    // The standby's FlushAcks advance the primary's slot to its watermark,
-    // so the primary can reclaim the shipped log.
-    wait_until("slot advance", || {
-        primary.storage().repl_slot_lsn(7) == Some(target)
-    })
-    .await;
-
-    // Reconnect: stop the receiver, commit more, restart it — the stream
-    // resumes from the standby's persisted watermark (the slot held the
-    // log in between).
-    rx_shutdown_tx.send(true).unwrap();
-    let _ = receiver_task.await;
-    for i in 31..=40 {
-        primary
-            .execute(&format!("INSERT INTO t VALUES ({i}, {i})"))
-            .expect("insert");
-    }
-    let (rx_shutdown_tx2, rx_shutdown_rx2) = watch::channel(false);
-    let receiver_task2 = tokio::spawn(run_standby_receiver(
-        receiver_cfg,
-        standby.storage_arc(),
-        rx_shutdown_rx2,
-    ));
-    let target = primary.storage().wal_flushed_lsn();
-    wait_until("resume after reconnect", || {
-        standby.storage().wal_tail() >= target
-    })
-    .await;
-    let expected = sql_rows(&primary, "SELECT id, v FROM t ORDER BY id").1;
-    assert_eq!(
-        sql_rows(&standby, "SELECT id, v FROM t ORDER BY id").1,
-        expected,
-        "the standby resumes and matches after a receiver restart"
-    );
-
-    rx_shutdown_tx2.send(true).unwrap();
-    let _ = receiver_task2.await;
-    shutdown_tx.send(true).unwrap();
-    let _ = listener_task.await;
-    drop(standby);
-    drop(primary);
-    for p in [primary_path, bak, standby_path] {
-        let _ = std::fs::remove_file(p);
-    }
-}
-
-// Stage 18 (18.3 restartpoints): a standby cannot checkpoint, so the
-// restartpoint is what reclaims its WAL ring — flush redone pages, persist
-// the allocator, advance the head to the standby's OWN undo floor (its
-// active-transaction table over the shipped log, plus the first search
-// record the seed snapshot does not cover). Storage-only: no WAL append,
-// no snapshot write (a locally allocated snapshot extent would collide
-// with the primary's future logged allocations).
-#[test]
-fn a_standby_restartpoint_reclaims_the_ring_up_to_its_own_undo_floor() {
-    let primary_path = unique_temp_path("rsp-primary");
-    let bak = unique_temp_path("rsp-bak");
-    let standby_path = unique_temp_path("rsp-standby");
-
-    // A pure-relational primary (no search events, so no search floor).
-    let primary = new_engine(&primary_path);
-    primary
-        .execute("CREATE TABLE t (id INT NOT NULL PRIMARY KEY, v INT)")
-        .expect("create");
-    for i in 1..=10 {
-        primary
-            .execute(&format!("INSERT INTO t VALUES ({i}, {i})"))
-            .expect("insert");
-    }
-    let summary = primary.storage().backup_full(&bak).expect("backup");
-    let backup_end = summary.backup_end_lsn;
-    for i in 11..=30 {
-        primary
-            .execute(&format!("INSERT INTO t VALUES ({i}, {i})"))
-            .expect("insert");
-    }
-    let flushed = primary.storage().wal_flushed_lsn();
-    let delta = primary
-        .storage()
-        .read_wal_range(backup_end, flushed)
-        .expect("delta");
-    let expected = sql_rows(&primary, "SELECT id, v FROM t ORDER BY id").1;
-
-    // A restartpoint on the primary is a no-op (not a standby).
-    assert!(
-        !primary
-            .standby_restartpoint_if_needed()
-            .expect("primary no-op"),
-        "a primary never takes a restartpoint"
-    );
-
-    // Every shipped transaction is resolved, so the floor is the applied
-    // tail: the restartpoint reclaims the whole shipped range.
-    Storage::restore_full_standby(&standby_path, &bak, &[]).expect("restore");
-    let standby = Engine::new(Storage::open(standby_path.clone()).expect("open")).expect("eng");
-    standby
-        .storage()
-        .apply_wal_stream(backup_end, &delta)
-        .expect("apply");
-    let head_before = standby.storage().wal_head();
-    assert!(
-        standby.standby_restartpoint().expect("restartpoint"),
-        "the restartpoint runs"
-    );
-    assert!(standby.storage().wal_head() > head_before);
-    assert_eq!(
-        standby.storage().wal_head(),
-        standby.storage().wal_tail(),
-        "with everything resolved, the head reaches the applied tail"
-    );
-    assert!(
-        !standby.standby_restartpoint().expect("second"),
-        "nothing further to reclaim"
-    );
-    assert_eq!(
-        sql_rows(&standby, "SELECT id, v FROM t ORDER BY id").1,
-        expected,
-        "relational state intact after the restartpoint"
-    );
-
-    // Durable: reopen replays from the advanced head (pages were flushed).
-    drop(standby);
-    let standby = Engine::new(Storage::open(standby_path.clone()).expect("reopen")).expect("eng");
-    assert_eq!(
-        sql_rows(&standby, "SELECT id, v FROM t ORDER BY id").1,
-        expected,
-        "state survives a reopen after the restartpoint"
-    );
-
-    // The stream continues past a restartpoint.
-    for i in 31..=40 {
-        primary
-            .execute(&format!("INSERT INTO t VALUES ({i}, {i})"))
-            .expect("insert");
-    }
-    let flushed2 = primary.storage().wal_flushed_lsn();
-    let delta2 = primary
-        .storage()
-        .read_wal_range(flushed, flushed2)
-        .expect("delta2");
-    standby
-        .storage()
-        .apply_wal_stream(flushed, &delta2)
-        .expect("apply after restartpoint");
-    let expected = sql_rows(&primary, "SELECT id, v FROM t ORDER BY id").1;
-    assert_eq!(
-        sql_rows(&standby, "SELECT id, v FROM t ORDER BY id").1,
-        expected,
-        "the stream continues cleanly after a restartpoint"
-    );
-
-    drop(primary);
-    drop(standby);
-    for p in [primary_path, bak, standby_path] {
-        let _ = std::fs::remove_file(p);
-    }
-}
-
-// A WRITABLE restore is a NEW timeline (its history rewinds and its future
-// writes diverge), so it must not present the source's epoch to standbys
-// that followed the source — the epoch bumps. A --standby seed stays on
-// the source's timeline verbatim.
-#[test]
-fn a_writable_restore_is_a_new_timeline_a_standby_seed_is_not() {
-    let src = unique_temp_path("tl-src");
-    let bak = unique_temp_path("tl-bak");
-    let writable = unique_temp_path("tl-writable");
-    let seed = unique_temp_path("tl-seed");
-
-    let engine = new_engine(&src);
-    engine
-        .execute("CREATE TABLE t (id INT NOT NULL PRIMARY KEY)")
-        .expect("create");
-    engine.execute("INSERT INTO t VALUES (1)").expect("insert");
-    engine.storage().backup_full(&bak).expect("backup");
-    drop(engine);
-
-    Storage::restore_full(&writable, &bak).expect("plain restore");
-    let restored = Engine::new(Storage::open(writable.clone()).expect("open")).expect("eng");
-    assert_eq!(
-        restored.storage().epoch(),
-        1,
-        "a writable restore bumps the epoch (new timeline)"
-    );
-    drop(restored);
-
-    Storage::restore_full_standby(&seed, &bak, &[]).expect("standby seed");
-    let standby = Engine::new(Storage::open(seed.clone()).expect("open")).expect("eng");
-    assert_eq!(
-        standby.storage().epoch(),
-        0,
-        "a standby seed carries the source's timeline"
-    );
-    drop(standby);
-    for p in [src, bak, writable, seed] {
-        let _ = std::fs::remove_file(p);
-    }
-}
-
-// `ALTER DATABASE <name> FAILOVER` online mirrors RESTORE DATABASE: the
-// answer is the pointer at the offline CLI.
-#[test]
-fn alter_database_failover_points_at_the_offline_cli() {
-    let path = unique_temp_path("failover-online");
-    let engine = new_engine(&path);
-    assert_eq!(
-        sql_error_number(&engine, "ALTER DATABASE CURRENT FAILOVER"),
-        3101
-    );
-    let response = sql(&engine, "ALTER DATABASE CURRENT FAILOVER");
-    assert!(
-        response["error"]["message"]
-            .as_str()
-            .unwrap_or_default()
-            .contains("truthdb-cli promote"),
-        "{response}"
-    );
-    drop(engine);
-    let _ = std::fs::remove_file(path);
-}
-
-// Stage 18.5 (D2) readable standby: reads resolve through the version
-// store at the last-applied-commit snapshot — shipped in-flight changes
-// (inserts AND updates of pre-existing rows) are invisible until their
-// commit ships; an aborted shipped transaction unwinds; the store is
-// rebuilt at reopen from the retained ring.
-#[test]
-fn a_readable_standby_serves_only_committed_state() {
-    let primary_path = unique_temp_path("read-primary");
-    let bak = unique_temp_path("read-bak");
-    let standby_path = unique_temp_path("read-standby");
-
-    let primary = new_engine(&primary_path);
-    let mut ctx = TxnContext::default();
-    batch(
-        &primary,
-        &mut ctx,
-        "CREATE TABLE t (id INT NOT NULL PRIMARY KEY, v INT)",
-    );
-    for i in 1..=10 {
-        batch(
-            &primary,
-            &mut ctx,
-            &format!("INSERT INTO t VALUES ({i}, {i})"),
-        );
-    }
-    let summary = primary.storage().backup_full(&bak).expect("backup");
-    let backup_end = summary.backup_end_lsn;
-    // Committed post-backup work...
-    for i in 11..=15 {
-        batch(
-            &primary,
-            &mut ctx,
-            &format!("INSERT INTO t VALUES ({i}, {i})"),
-        );
-    }
-    // ...then an in-flight transaction: new rows AND an update of row 1.
-    batch(&primary, &mut ctx, "BEGIN TRANSACTION");
-    for i in 16..=20 {
-        batch(
-            &primary,
-            &mut ctx,
-            &format!("INSERT INTO t VALUES ({i}, {i})"),
-        );
-    }
-    batch(&primary, &mut ctx, "UPDATE t SET v = 999 WHERE id = 1");
-    let mid = primary.storage().wal_flushed_lsn();
-    let delta = primary
-        .storage()
-        .read_wal_range(backup_end, mid)
-        .expect("delta");
-
-    Storage::restore_full_standby(&standby_path, &bak, &[]).expect("seed");
-    let standby = Engine::new(Storage::open(standby_path.clone()).expect("open")).expect("eng");
-    standby
-        .storage()
-        .apply_wal_stream(backup_end, &delta)
-        .expect("apply");
-    assert_eq!(
-        sql_rows(&standby, "SELECT COUNT(*) FROM t").1,
-        vec![vec![Some("15".into())]],
-        "in-flight inserts are invisible; committed ones are visible"
-    );
-    assert_eq!(
-        sql_rows(&standby, "SELECT v FROM t WHERE id = 1").1,
-        vec![vec![Some("1".into())]],
-        "an in-flight update serves the committed pre-image"
-    );
-
-    // The store rebuilds at reopen from the retained ring.
-    drop(standby);
-    let standby = Engine::new(Storage::open(standby_path.clone()).expect("open")).expect("eng");
-    assert_eq!(
-        sql_rows(&standby, "SELECT v FROM t WHERE id = 1").1,
-        vec![vec![Some("1".into())]],
-        "committed-state reads survive a reopen"
-    );
-    assert_eq!(
-        sql_rows(&standby, "SELECT COUNT(*) FROM t").1,
-        vec![vec![Some("15".into())]]
-    );
-
-    // The commit ships: everything becomes visible.
-    batch(&primary, &mut ctx, "COMMIT TRANSACTION");
-    let committed = primary.storage().wal_flushed_lsn();
-    let delta2 = primary
-        .storage()
-        .read_wal_range(mid, committed)
-        .expect("commit range");
-    standby
-        .storage()
-        .apply_wal_stream(mid, &delta2)
-        .expect("apply commit");
-    assert_eq!(
-        sql_rows(&standby, "SELECT COUNT(*) FROM t").1,
-        vec![vec![Some("20".into())]],
-        "the committed transaction is visible"
-    );
-    assert_eq!(
-        sql_rows(&standby, "SELECT v FROM t WHERE id = 1").1,
-        vec![vec![Some("999".into())]],
-        "the committed update is visible"
-    );
-
-    // An aborted shipped transaction: its effects never surface, and its
-    // version chains unwind.
-    batch(&primary, &mut ctx, "BEGIN TRANSACTION");
-    batch(&primary, &mut ctx, "UPDATE t SET v = -1 WHERE id = 2");
-    batch(&primary, &mut ctx, "ROLLBACK TRANSACTION");
-    let aborted = primary.storage().wal_flushed_lsn();
-    let delta3 = primary
-        .storage()
-        .read_wal_range(committed, aborted)
-        .expect("abort range");
-    standby
-        .storage()
-        .apply_wal_stream(committed, &delta3)
-        .expect("apply abort");
-    assert_eq!(
-        sql_rows(&standby, "SELECT v FROM t WHERE id = 2").1,
-        vec![vec![Some("2".into())]],
-        "an aborted shipped transaction never surfaces"
-    );
-
-    drop(primary);
-    drop(standby);
-    for p in [primary_path, bak, standby_path] {
-        let _ = std::fs::remove_file(p);
-    }
-}
-
-// The two review-caught wrong-results bugs, pinned: heap undo payloads
-// are CELL bytes (a tag byte before the row — served raw they decode as
-// garbage), and a statement rollback inside a transaction that LATER
-// COMMITS ships CLRs with no TXN_END — un-captured, the rolled-back
-// delete's chain head would hide a committed, physically present row.
-#[test]
-fn readable_standby_handles_heap_cells_and_statement_rollbacks() {
-    let primary_path = unique_temp_path("edge-primary");
-    let bak = unique_temp_path("edge-bak");
-    let standby_path = unique_temp_path("edge-standby");
-
-    let primary = new_engine(&primary_path);
-    let mut ctx = TxnContext::default();
-    // A HEAP table (no primary key) — its undo payloads are tagged cells.
-    batch(&primary, &mut ctx, "CREATE TABLE h (id INT, v INT)");
-    for i in 1..=3 {
-        batch(
-            &primary,
-            &mut ctx,
-            &format!("INSERT INTO h VALUES ({i}, {i})"),
-        );
-    }
-    // A keyed table for the savepoint flow.
-    batch(
-        &primary,
-        &mut ctx,
-        "CREATE TABLE t (id INT NOT NULL PRIMARY KEY, v INT)",
-    );
-    for i in 1..=3 {
-        batch(
-            &primary,
-            &mut ctx,
-            &format!("INSERT INTO t VALUES ({i}, {i})"),
-        );
-    }
-    let summary = primary.storage().backup_full(&bak).expect("backup");
-    let backup_end = summary.backup_end_lsn;
-
-    // In-flight heap update: the standby must serve the OLD row, decoded
-    // correctly (the tag byte stripped).
-    batch(&primary, &mut ctx, "BEGIN TRANSACTION");
-    batch(&primary, &mut ctx, "UPDATE h SET v = 999 WHERE id = 1");
-    let mid = primary.storage().wal_flushed_lsn();
-    let delta = primary
-        .storage()
-        .read_wal_range(backup_end, mid)
-        .expect("delta");
-
-    Storage::restore_full_standby(&standby_path, &bak, &[]).expect("seed");
-    let standby = Engine::new(Storage::open(standby_path.clone()).expect("open")).expect("eng");
-    standby
-        .storage()
-        .apply_wal_stream(backup_end, &delta)
-        .expect("apply");
-    assert_eq!(
-        sql_rows(&standby, "SELECT id, v FROM h ORDER BY id").1,
-        vec![
-            vec![Some("1".into()), Some("1".into())],
-            vec![Some("2".into()), Some("2".into())],
-            vec![Some("3".into()), Some("3".into())],
-        ],
-        "an in-flight heap update serves the committed pre-image, decoded as a row"
-    );
-
-    // Commit the heap txn, then a mid-transaction STATEMENT rollback: a
-    // multi-row re-key that collides part-way is undone with CLRs (no
-    // TXN_END — the transaction continues and commits). Un-captured, the
-    // undone delete's chain head would hide committed row 1 forever.
-    batch(&primary, &mut ctx, "COMMIT TRANSACTION");
-    batch(&primary, &mut ctx, "INSERT INTO t VALUES (10, 10)");
-    batch(&primary, &mut ctx, "BEGIN TRANSACTION");
-    let failed = batch(&primary, &mut ctx, "UPDATE t SET id = id + 9 WHERE id <= 2");
-    assert!(
-        failed.error.is_some(),
-        "the re-key collides with row 10 and the statement rolls back"
-    );
-    batch(&primary, &mut ctx, "INSERT INTO t VALUES (4, 4)");
-    batch(&primary, &mut ctx, "COMMIT TRANSACTION");
-    assert_eq!(
-        sql_rows(&primary, "SELECT COUNT(*) FROM t").1,
-        vec![vec![Some("5".into())]],
-        "primary: rows 1,2,3,10,4 (the failed statement fully undone)"
-    );
-    let done = primary.storage().wal_flushed_lsn();
-    let delta2 = primary.storage().read_wal_range(mid, done).expect("delta2");
-    standby
-        .storage()
-        .apply_wal_stream(mid, &delta2)
-        .expect("apply2");
-    assert_eq!(
-        sql_rows(&standby, "SELECT id FROM t ORDER BY id").1,
-        sql_rows(&primary, "SELECT id FROM t ORDER BY id").1,
-        "the CLR-undone statement's rows stay visible on the standby"
-    );
-    assert_eq!(
-        sql_rows(&standby, "SELECT v FROM t WHERE id = 1").1,
-        vec![vec![Some("1".into())]],
-        "row 1 — deleted, CLR-restored, then committed — is served"
-    );
-    assert_eq!(
-        sql_rows(&standby, "SELECT v FROM h WHERE id = 1").1,
-        vec![vec![Some("999".into())]],
-        "the committed heap update is visible"
-    );
-
-    drop(primary);
-    drop(standby);
-    for p in [primary_path, bak, standby_path] {
-        let _ = std::fs::remove_file(p);
-    }
-}
-
-// Stage 18.6 monitoring: the replication DMVs report role, slots,
-// connectedness, and lag on both sides of a link.
-#[test]
-fn replication_dmvs_report_slots_lag_and_role() {
-    let primary_path = unique_temp_path("dmv-primary");
-    let bak = unique_temp_path("dmv-bak");
-    let standby_path = unique_temp_path("dmv-standby");
-
-    let primary = new_engine(&primary_path);
-    primary
-        .execute("CREATE TABLE t (id INT NOT NULL PRIMARY KEY)")
-        .expect("create");
-    for i in 1..=5 {
-        primary
-            .execute(&format!("INSERT INTO t VALUES ({i})"))
-            .expect("insert");
-    }
-    // No slots yet: the primary reports zero replica rows.
-    assert_eq!(
-        sql_rows(&primary, "SELECT COUNT(*) FROM sys.dm_repl_replica_states").1,
-        vec![vec![Some("0".into())]],
-    );
-    assert_eq!(
-        sql_rows(&primary, "SELECT COUNT(*) FROM sys.dm_repl_slots").1[0][0],
-        Some("0".into())
-    );
-
-    // Register a slot (as the sender would) and check the reporting.
-    let held = primary.storage().wal_flushed_lsn();
-    primary
-        .storage()
-        .try_register_repl_slot(3, held)
-        .expect("slot");
-    for i in 6..=10 {
-        primary
-            .execute(&format!("INSERT INTO t VALUES ({i})"))
-            .expect("insert");
-    }
-    let (cols, rows) = sql_rows(
-        &primary,
-        "SELECT role, node_id, is_connected, lag_bytes, sync_state \
-             FROM sys.dm_repl_replica_states",
-    );
-    assert_eq!(
-        cols,
-        vec!["role", "node_id", "is_connected", "lag_bytes", "sync_state"]
-    );
-    assert_eq!(rows.len(), 1);
-    assert_eq!(rows[0][0], Some("PRIMARY".into()));
-    assert_eq!(rows[0][1], Some("3".into()));
-    assert_eq!(
-        rows[0][2],
-        Some("0".into()),
-        "no live sender (bit renders 0)"
-    );
-    let lag: i64 = rows[0][3].as_deref().unwrap().parse().expect("lag");
-    assert!(lag > 0, "the slot lags the new commits");
-    assert_eq!(rows[0][4], Some("ASYNC".into()));
-    let (_, slot_rows) = sql_rows(
-        &primary,
-        "SELECT slot_id, retained_bytes FROM sys.dm_repl_slots",
-    );
-    assert_eq!(slot_rows[0][0], Some("3".into()));
-    assert!(slot_rows[0][1].as_deref().unwrap().parse::<i64>().unwrap() > 0);
-
-    // A standby reports its own role and applied position.
-    let summary = primary.storage().backup_full(&bak).expect("backup");
-    Storage::restore_full_standby(&standby_path, &bak, &[]).expect("seed");
-    let standby = Engine::new(Storage::open(standby_path.clone()).expect("open")).expect("eng");
-    let _ = summary;
-    let (_, srows) = sql_rows(
-        &standby,
-        "SELECT role, sync_state FROM sys.dm_repl_replica_states",
-    );
-    assert_eq!(srows.len(), 1);
-    assert_eq!(srows[0][0], Some("STANDBY".into()));
-    assert_eq!(srows[0][1], Some("NOT_APPLICABLE".into()));
-
-    drop(primary);
-    drop(standby);
-    for p in [primary_path, bak, standby_path] {
-        let _ = std::fs::remove_file(p);
-    }
-}
-
-// Stage 18.5 (D2) synchronous commit: a commit waits — after local
-// durability — for a standby acknowledgement; a timeout degrades the link
-// (NOT_SYNCHRONIZED) availability-first, and a caught-up acknowledgement
-// re-synchronizes it.
-#[test]
-fn synchronous_commit_waits_for_acks_and_degrades_on_timeout() {
-    let path = unique_temp_path("sync-commit");
-    let engine = new_engine(&path);
-    engine
-        .execute("CREATE TABLE t (id INT NOT NULL PRIMARY KEY)")
-        .expect("create");
-    engine
-        .storage()
-        .arm_sync_commit(std::time::Duration::from_millis(500));
-
-    // A healthy standby: an acker thread mirrors the durable watermark.
-    let storage = engine.storage_arc();
-    let stop = std::sync::Arc::new(std::sync::atomic::AtomicBool::new(false));
-    let stop2 = std::sync::Arc::clone(&stop);
-    let acker = std::thread::spawn(move || {
-        while !stop2.load(std::sync::atomic::Ordering::Relaxed) {
-            storage.publish_sync_ack(storage.wal_flushed_lsn());
-            std::thread::sleep(std::time::Duration::from_millis(5));
-        }
-    });
-    engine.execute("INSERT INTO t VALUES (1)").expect("insert");
-    assert!(
-        !engine.storage().sync_commit_degraded(),
-        "an acked commit keeps the link synchronized"
-    );
-    stop.store(true, std::sync::atomic::Ordering::Relaxed);
-    acker.join().expect("acker");
-
-    // The standby goes silent: the next commit stalls out the timeout,
-    // degrades the link, and still succeeds.
-    let t0 = std::time::Instant::now();
-    engine.execute("INSERT INTO t VALUES (2)").expect("insert");
-    assert!(
-        t0.elapsed() >= std::time::Duration::from_millis(500),
-        "the degrading commit waited out the timeout"
-    );
-    assert!(
-        engine.storage().sync_commit_degraded(),
-        "the timed-out link is NOT_SYNCHRONIZED"
-    );
-
-    // Degraded: commits pass straight through (bounded by well under the
-    // timeout — generous margin against CI jitter).
-    let t1 = std::time::Instant::now();
-    engine.execute("INSERT INTO t VALUES (3)").expect("insert");
-    assert!(
-        t1.elapsed() < std::time::Duration::from_millis(400),
-        "a degraded link does not delay commits"
-    );
-
-    // A caught-up acknowledgement re-synchronizes.
-    engine
-        .storage()
-        .publish_sync_ack(engine.storage().wal_flushed_lsn());
-    assert!(
-        !engine.storage().sync_commit_degraded(),
-        "a caught-up standby re-synchronizes the link"
-    );
-    drop(engine);
-    let _ = std::fs::remove_file(path);
-}
-
-// Stage 18.4 (D1 manual failover): offline promotion turns a standby into
-// a read-write primary — finish recovery (undo the shipped in-flight
-// transactions), clear the standby mode, bump the epoch. The epoch fences
-// the old timeline: only an EQUAL epoch may follow the new primary, so
-// the old primary and pre-failover seeds must reseed.
-#[test]
-fn promotion_finishes_recovery_bumps_the_epoch_and_fences_the_old_timeline() {
-    use crate::repl::handshake::{HandshakeParams, compute_auth, evaluate_hello};
-
-    let primary_path = unique_temp_path("promo-primary");
-    let bak = unique_temp_path("promo-bak");
-    let standby_path = unique_temp_path("promo-standby");
-    let reseed_bak = unique_temp_path("promo-reseed-bak");
-    let reseed_path = unique_temp_path("promo-reseed");
-
-    let primary = new_engine(&primary_path);
-    let mut ctx = TxnContext::default();
-    batch(
-        &primary,
-        &mut ctx,
-        "CREATE TABLE t (id INT NOT NULL PRIMARY KEY, v INT)",
-    );
-    for i in 1..=10 {
-        batch(
-            &primary,
-            &mut ctx,
-            &format!("INSERT INTO t VALUES ({i}, {i})"),
-        );
-    }
-    let summary = primary.storage().backup_full(&bak).expect("backup");
-    let backup_end = summary.backup_end_lsn;
-    for i in 11..=20 {
-        batch(
-            &primary,
-            &mut ctx,
-            &format!("INSERT INTO t VALUES ({i}, {i})"),
-        );
-    }
-    // An in-flight transaction, durable but uncommitted: the primary dies
-    // here; promotion must roll it back.
-    batch(&primary, &mut ctx, "BEGIN TRANSACTION");
-    for i in 21..=25 {
-        batch(
-            &primary,
-            &mut ctx,
-            &format!("INSERT INTO t VALUES ({i}, {i})"),
-        );
-    }
-    let mid = primary.storage().wal_flushed_lsn();
-    let delta = primary
-        .storage()
-        .read_wal_range(backup_end, mid)
-        .expect("delta");
-
-    Storage::restore_full_standby(&standby_path, &bak, &[]).expect("restore");
-    let standby = Engine::new(Storage::open(standby_path.clone()).expect("open")).expect("eng");
-    standby
-        .storage()
-        .apply_wal_stream(backup_end, &delta)
-        .expect("apply");
-
-    // A running server holds the advisory lock: promote refuses.
-    assert!(
-        Storage::promote(&standby_path).is_err(),
-        "promote refuses a file a live server holds"
-    );
-    drop(standby);
-
-    // A non-standby cannot be promoted.
-    drop(primary);
-    let err = Storage::promote(&primary_path).expect_err("primary is not a standby");
-    assert!(
-        err.to_string().contains("not a replication standby"),
-        "{err}"
-    );
-
-    let epoch = Storage::promote(&standby_path).expect("promote");
-    assert_eq!(epoch, 1, "the first failover bumps epoch 0 -> 1");
-
-    let promoted = Engine::new(Storage::open(standby_path.clone()).expect("open")).expect("eng");
-    assert!(
-        !promoted.storage().is_standby(),
-        "the promoted node is a normal primary"
-    );
-    assert_eq!(promoted.storage().epoch(), 1);
-    assert_eq!(
-        sql_rows(&promoted, "SELECT COUNT(*) FROM t").1,
-        vec![vec![Some("20".into())]],
-        "the shipped in-flight transaction was undone at promotion"
-    );
-    promoted
-        .execute("INSERT INTO t VALUES (999, 999)")
-        .expect("the promoted node accepts writes");
-
-    // Epoch fencing: the OLD timeline (epoch 0 — the dead primary, or any
-    // pre-failover seed) cannot follow the new primary.
-    const SECRET: &[u8] = b"promo-secret";
-    const UUID: [u8; 16] = [9u8; 16];
-    let params = HandshakeParams {
-        shared_secret: SECRET,
-        cluster_uuid: UUID,
-        primary_epoch: promoted.storage().epoch(),
-        primary_flushed_lsn: promoted.storage().wal_flushed_lsn(),
-    };
-    let old_auth = compute_auth(SECRET, 3, &UUID, 0, mid);
-    let old_hello = crate::repl::Hello {
-        protocol_version: crate::repl::REPL_PROTOCOL_VERSION,
-        node_id: 3,
-        cluster_uuid: UUID,
-        epoch: 0,
-        last_received_lsn: mid,
-        auth: old_auth,
-    };
-    let ack = evaluate_hello(&old_hello, &params);
-    assert!(!ack.accepted, "the old timeline is fenced off");
-    assert!(ack.message.contains("reseed"), "{}", ack.message);
-
-    // A FRESH seed from the new primary carries epoch 1 and is accepted.
-    promoted
-        .storage()
-        .backup_full(&reseed_bak)
-        .expect("backup of the promoted node");
-    Storage::restore_full_standby(&reseed_path, &reseed_bak, &[]).expect("reseed");
-    let reseeded = Engine::new(Storage::open(reseed_path.clone()).expect("open")).expect("eng");
-    assert_eq!(
-        reseeded.storage().epoch(),
-        1,
-        "the seed carries the new timeline's epoch"
-    );
-    let new_auth = compute_auth(SECRET, 3, &UUID, 1, reseeded.storage().wal_tail());
-    let new_hello = crate::repl::Hello {
-        protocol_version: crate::repl::REPL_PROTOCOL_VERSION,
-        node_id: 3,
-        cluster_uuid: UUID,
-        epoch: 1,
-        last_received_lsn: reseeded.storage().wal_tail(),
-        auth: new_auth,
-    };
-    assert!(
-        evaluate_hello(&new_hello, &params).accepted,
-        "an equal-epoch seed follows the new primary"
-    );
-
-    drop(promoted);
-    drop(reseeded);
-    for p in [primary_path, bak, standby_path, reseed_bak, reseed_path] {
-        let _ = std::fs::remove_file(p);
-    }
-}
-
-// A FAILED apply (a cut or corrupt chunk) advances the live ring tail but
-// not the persisted one, and folds nothing into the restartpoint floors.
-// The restartpoint must reclaim only up to the persisted tail — the last
-// fully decoded + redone + committed range — never over bytes whose redo
-// did not run.
-#[test]
-fn a_failed_apply_never_feeds_a_restartpoint() {
-    let primary_path = unique_temp_path("rspx-primary");
-    let bak = unique_temp_path("rspx-bak");
-    let standby_path = unique_temp_path("rspx-standby");
-
-    let primary = new_engine(&primary_path);
-    primary
-        .execute("CREATE TABLE t (id INT NOT NULL PRIMARY KEY, v INT)")
-        .expect("create");
-    for i in 1..=10 {
-        primary
-            .execute(&format!("INSERT INTO t VALUES ({i}, {i})"))
-            .expect("insert");
-    }
-    let summary = primary.storage().backup_full(&bak).expect("backup");
-    let backup_end = summary.backup_end_lsn;
-    for i in 11..=20 {
-        primary
-            .execute(&format!("INSERT INTO t VALUES ({i}, {i})"))
-            .expect("insert");
-    }
-    let f1 = primary.storage().wal_flushed_lsn();
-    let delta1 = primary
-        .storage()
-        .read_wal_range(backup_end, f1)
-        .expect("delta1");
-    for i in 21..=30 {
-        primary
-            .execute(&format!("INSERT INTO t VALUES ({i}, {i})"))
-            .expect("insert");
-    }
-    let f2 = primary.storage().wal_flushed_lsn();
-    let delta2 = primary.storage().read_wal_range(f1, f2).expect("delta2");
-
-    Storage::restore_full_standby(&standby_path, &bak, &[]).expect("restore");
-    let standby = Engine::new(Storage::open(standby_path.clone()).expect("open")).expect("eng");
-    standby
-        .storage()
-        .apply_wal_stream(backup_end, &delta1)
-        .expect("apply delta1");
-
-    // A cut chunk fails the apply's coverage guard AFTER the ring bytes
-    // were seeded (the live tail moved; the persisted tail did not).
-    let cut = &delta2[..delta2.len() - 3];
-    standby
-        .storage()
-        .apply_wal_stream(f1, cut)
-        .expect_err("a cut chunk is refused");
-
-    // The restartpoint reclaims only to the persisted tail.
-    assert!(
-        standby.standby_restartpoint().expect("restartpoint"),
-        "the fully-applied prefix is reclaimable"
-    );
-    assert_eq!(
-        standby.storage().wal_head(),
-        f1,
-        "the head stops at the persisted tail, not at the failed chunk's end"
-    );
-
-    // The re-shipped full chunk applies cleanly over the cut bytes, and
-    // the next restartpoint reaches the new tail.
-    standby
-        .storage()
-        .apply_wal_stream(f1, &delta2)
-        .expect("re-apply the full chunk");
-    assert!(
-        standby.standby_restartpoint().expect("after re-apply"),
-        "the re-applied range is reclaimable"
-    );
-    assert_eq!(standby.storage().wal_head(), f2);
-    let expected = sql_rows(&primary, "SELECT id, v FROM t ORDER BY id").1;
-    assert_eq!(
-        sql_rows(&standby, "SELECT id, v FROM t ORDER BY id").1,
-        expected,
-        "the standby matches the primary after the recovery"
-    );
-
-    drop(primary);
-    drop(standby);
-    for p in [primary_path, bak, standby_path] {
-        let _ = std::fs::remove_file(p);
-    }
-}
-
-// The restartpoint's undo floor: a shipped range ending MID-transaction
-// leaves that transaction unresolved in the standby's own
-// active-transaction table, and the head must stop at its BEGIN — the
-// undo log above it is what promotion's analysis+undo will need. Once the
-// commit ships, the floor lifts.
-#[test]
-fn a_standby_restartpoint_stops_at_an_unresolved_transactions_begin() {
-    let primary_path = unique_temp_path("rspf-primary");
-    let bak = unique_temp_path("rspf-bak");
-    let standby_path = unique_temp_path("rspf-standby");
-
-    let primary = new_engine(&primary_path);
-    let mut ctx = TxnContext::default();
-    batch(
-        &primary,
-        &mut ctx,
-        "CREATE TABLE t (id INT NOT NULL PRIMARY KEY, v INT)",
-    );
-    for i in 1..=10 {
-        batch(
-            &primary,
-            &mut ctx,
-            &format!("INSERT INTO t VALUES ({i}, {i})"),
-        );
-    }
-    let summary = primary.storage().backup_full(&bak).expect("backup");
-    let backup_end = summary.backup_end_lsn;
-    let begin_area = primary.storage().wal_flushed_lsn();
-
-    // An explicit transaction, durable but uncommitted: the shipped range
-    // ends mid-transaction.
-    batch(&primary, &mut ctx, "BEGIN TRANSACTION");
-    for i in 11..=20 {
-        batch(
-            &primary,
-            &mut ctx,
-            &format!("INSERT INTO t VALUES ({i}, {i})"),
-        );
-    }
-    let mid = primary.storage().wal_flushed_lsn();
-    let delta = primary
-        .storage()
-        .read_wal_range(backup_end, mid)
-        .expect("mid-transaction range");
-
-    Storage::restore_full_standby(&standby_path, &bak, &[]).expect("restore");
-    let standby = Engine::new(Storage::open(standby_path.clone()).expect("open")).expect("eng");
-    standby
-        .storage()
-        .apply_wal_stream(backup_end, &delta)
-        .expect("apply mid-txn range");
-    assert!(
-        standby.standby_restartpoint().expect("restartpoint"),
-        "the resolved prefix is reclaimable"
-    );
-    let head = standby.storage().wal_head();
-    assert_eq!(
-        head, begin_area,
-        "the head advanced exactly to the unresolved transaction's BEGIN \
-             (reclaiming everything before it, retaining all of its undo log)"
-    );
-    assert!(
-        head < mid,
-        "the head stopped below the applied tail (the open transaction pins it)"
-    );
-
-    // The floor survives a reopen (the ATT is re-derived from the ring).
-    drop(standby);
-    let standby = Engine::new(Storage::open(standby_path.clone()).expect("open")).expect("eng");
-    assert!(
-        !standby.standby_restartpoint().expect("after reopen"),
-        "still pinned by the same unresolved transaction after a reopen"
-    );
-
-    // The commit ships: the floor lifts and the next restartpoint reaches
-    // the applied tail.
-    batch(&primary, &mut ctx, "COMMIT");
-    let flushed = primary.storage().wal_flushed_lsn();
-    let delta2 = primary
-        .storage()
-        .read_wal_range(mid, flushed)
-        .expect("commit range");
-    standby
-        .storage()
-        .apply_wal_stream(mid, &delta2)
-        .expect("apply commit");
-    assert!(
-        standby.standby_restartpoint().expect("after commit"),
-        "the commit lifted the floor"
-    );
-    assert_eq!(
-        standby.storage().wal_head(),
-        standby.storage().wal_tail(),
-        "everything resolved: full reclaim"
-    );
-    let expected = sql_rows(&primary, "SELECT id, v FROM t ORDER BY id").1;
-    assert_eq!(
-        sql_rows(&standby, "SELECT id, v FROM t ORDER BY id").1,
-        expected,
-        "the standby matches the primary"
-    );
-
-    drop(primary);
-    drop(standby);
-    for p in [primary_path, bak, standby_path] {
-        let _ = std::fs::remove_file(p);
-    }
-}
-
-// The restartpoint's search floor: the seed carries no search snapshot, so
-// every shipped search record must stay in the ring for the reopen replay
-// — the head clamps below the first one, and the reopened standby still
-// serves the search state.
-#[test]
-fn a_standby_restartpoint_never_truncates_unsnapshotted_search_records() {
-    let primary_path = unique_temp_path("rsps-primary");
-    let bak = unique_temp_path("rsps-bak");
-    let standby_path = unique_temp_path("rsps-standby");
-
-    let primary = new_engine(&primary_path);
-    primary
-        .execute("CREATE TABLE t (id INT NOT NULL PRIMARY KEY)")
-        .expect("create");
-    for i in 1..=10 {
-        primary
-            .execute(&format!("INSERT INTO t VALUES ({i})"))
-            .expect("insert");
-    }
-    let summary = primary.storage().backup_full(&bak).expect("backup");
-    let backup_end = summary.backup_end_lsn;
-    // Post-backup: relational rows FIRST (a reclaimable prefix), then the
-    // search records (which pin the head), then more rows.
-    for i in 11..=15 {
-        primary
-            .execute(&format!("INSERT INTO t VALUES ({i})"))
-            .expect("insert");
-    }
-    primary
-            .execute(r#"create index products { "mappings": { "properties": { "name": { "type": "text" } } } }"#)
-            .expect("create search index");
-    primary
-        .execute(r#"insert document products { "name": "red shoes" }"#)
-        .expect("insert doc");
-    for i in 16..=20 {
-        primary
-            .execute(&format!("INSERT INTO t VALUES ({i})"))
-            .expect("insert");
-    }
-    let flushed = primary.storage().wal_flushed_lsn();
-    let delta = primary
-        .storage()
-        .read_wal_range(backup_end, flushed)
-        .expect("delta");
-
-    Storage::restore_full_standby(&standby_path, &bak, &[]).expect("restore");
-    let standby = Engine::new(Storage::open(standby_path.clone()).expect("open")).expect("eng");
-    standby
-        .storage()
-        .apply_wal_stream(backup_end, &delta)
-        .expect("apply");
-    assert!(
-        standby.standby_restartpoint().expect("restartpoint"),
-        "the pre-search prefix is reclaimable"
-    );
-    let head = standby.storage().wal_head();
-    assert!(
-        head > backup_end,
-        "the relational prefix before the search records was reclaimed"
-    );
-    assert!(
-        head < standby.storage().wal_tail(),
-        "the head stopped below the tail: the search records are retained"
-    );
-
-    // Reopen: the ring still holds every search record, so the replayed
-    // search state serves the streamed doc.
-    drop(standby);
-    let standby = Engine::new(Storage::open(standby_path.clone()).expect("open")).expect("eng");
-    let response = standby
-        .execute(r#"search products { "query": { "match": { "name": "red" } } }"#)
-        .expect("standby search after reopen");
-    let response: Value = serde_json::from_str(&response).expect("json");
-    assert_eq!(
-        response["hits"]["total"].as_u64(),
-        Some(1),
-        "the streamed search doc survives the restartpoint + reopen"
-    );
-
-    drop(primary);
-    drop(standby);
-    for p in [primary_path, bak, standby_path] {
-        let _ = std::fs::remove_file(p);
-    }
-}
-
-// A FULL-recovery-model primary's backup seeds the standby with the FULL
-// bit and a log-backup marker — but the standby must NOT hold ring
-// truncation there (its log chain belongs to the primary; holding would
-// cap every restartpoint at the seed point forever and run the ring
-// full). BACKUP LOG on the standby is refused for the same reason.
-#[test]
-fn a_full_model_standby_still_reclaims_and_refuses_backup_log() {
-    let primary_path = unique_temp_path("rspl-primary");
-    let bak = unique_temp_path("rspl-bak");
-    let standby_path = unique_temp_path("rspl-standby");
-
-    let primary = new_engine(&primary_path);
-    primary
-        .execute("CREATE TABLE t (id INT NOT NULL PRIMARY KEY)")
-        .expect("create");
-    primary
-        .execute("ALTER DATABASE truthdb SET RECOVERY FULL")
-        .expect("full model");
-    for i in 1..=10 {
-        primary
-            .execute(&format!("INSERT INTO t VALUES ({i})"))
-            .expect("insert");
-    }
-    let summary = primary.storage().backup_full(&bak).expect("backup");
-    let backup_end = summary.backup_end_lsn;
-    for i in 11..=20 {
-        primary
-            .execute(&format!("INSERT INTO t VALUES ({i})"))
-            .expect("insert");
-    }
-    let flushed = primary.storage().wal_flushed_lsn();
-    let delta = primary
-        .storage()
-        .read_wal_range(backup_end, flushed)
-        .expect("delta");
-
-    Storage::restore_full_standby(&standby_path, &bak, &[]).expect("restore");
-    let standby = Engine::new(Storage::open(standby_path.clone()).expect("open")).expect("eng");
-    standby
-        .storage()
-        .apply_wal_stream(backup_end, &delta)
-        .expect("apply");
-    assert!(
-        standby.standby_restartpoint().expect("restartpoint"),
-        "a FULL-model standby still reclaims (no log-backup hold)"
-    );
-    assert!(
-        standby.storage().wal_head() > backup_end,
-        "the head advanced past the seeded log-backup marker"
-    );
-    let err = sql(&standby, "BACKUP LOG truthdb TO DISK = '/tmp/never.trn'");
-    assert!(
-        !err["error"].is_null(),
-        "BACKUP LOG on a standby is refused: {err}"
-    );
-
-    drop(primary);
-    drop(standby);
-    for p in [primary_path, bak, standby_path] {
-        let _ = std::fs::remove_file(p);
-    }
-}
-
-// Stage 18 slice 4d (review fix): a shipped stream can end MID-transaction
-// (the primary's durable watermark can land between an in-flight txn's page
-// ops and its commit). A standby's reopen must REPEAT history (redo-only) —
-// a full ARIES undo would roll back the in-flight rows, and since the primary
-// commits them and resumes shipping ABOVE the standby's applied point, they
-// would be lost forever (silent replica divergence).
-#[test]
-fn a_standby_reopen_is_redo_only_and_keeps_in_flight_stream_data() {
-    let primary_path = unique_temp_path("redo-primary");
-    let bak = unique_temp_path("redo-bak");
-    let standby_path = unique_temp_path("redo-standby");
-
-    let primary = new_engine(&primary_path);
-    let mut ctx = TxnContext::default();
-    batch(
-        &primary,
-        &mut ctx,
-        "CREATE TABLE t (id INT NOT NULL PRIMARY KEY, v INT)",
-    );
-    for i in 1..=10 {
-        batch(
-            &primary,
-            &mut ctx,
-            &format!("INSERT INTO t VALUES ({i}, {i})"),
-        );
-    }
-    let summary = primary.storage().backup_full(&bak).expect("backup");
-    let backup_end = summary.backup_end_lsn;
-
-    // An explicit transaction whose inserts are made durable (fsynced) but NOT
-    // committed — the durable watermark now lands mid-transaction.
-    batch(&primary, &mut ctx, "BEGIN TRANSACTION");
-    for i in 11..=20 {
-        batch(
-            &primary,
-            &mut ctx,
-            &format!("INSERT INTO t VALUES ({i}, {i})"),
-        );
-    }
-    let mid = primary.storage().wal_flushed_lsn();
-    let delta = primary
-        .storage()
-        .read_wal_range(backup_end, mid)
-        .expect("ship the mid-transaction range");
-
-    let count = |eng: &Engine| -> String {
-        sql_rows(eng, "SELECT COUNT(*) FROM t").1[0][0]
-            .clone()
-            .unwrap_or_default()
-    };
-
-    // Standby applies the mid-transaction stream live: the in-flight rows
-    // are applied PHYSICALLY via redo (proven by the post-commit equality
-    // below) but hidden from reads — a readable standby serves only
-    // committed state.
-    Storage::restore_full(&standby_path, &bak).expect("restore");
-    {
-        let standby = Engine::new(Storage::open(standby_path.clone()).expect("open")).expect("eng");
-        standby
-            .storage()
-            .apply_wal_stream(backup_end, &delta)
-            .expect("apply mid-txn stream");
-        assert_eq!(
-            count(&standby),
-            "10",
-            "in-flight rows are applied but not visible"
-        );
-    }
-    // THE FIX: a standby reopen is redo-only, so the applied rows survive. A
-    // full ARIES undo here would drop the 10 in-flight rows permanently.
-    let standby = Engine::new(Storage::open(standby_path.clone()).expect("reopen")).expect("eng");
-    assert_eq!(
-        count(&standby),
-        "10",
-        "redo-only reopen keeps the streamed in-flight rows physically, still hidden"
-    );
-
-    // The primary commits and ships the continuation; the standby stays
-    // consistent with the primary.
-    batch(&primary, &mut ctx, "COMMIT TRANSACTION");
-    let after = primary.storage().wal_flushed_lsn();
-    let delta2 = primary
-        .storage()
-        .read_wal_range(mid, after)
-        .expect("ship the commit");
-    standby
-        .storage()
-        .apply_wal_stream(mid, &delta2)
-        .expect("apply the commit");
-    assert_eq!(
-        sql_rows(&standby, "SELECT id, v FROM t ORDER BY id").1,
-        sql_rows(&primary, "SELECT id, v FROM t ORDER BY id").1,
-        "the standby matches the primary once the commit is shipped"
-    );
-
-    drop(primary);
-    drop(standby);
-    for p in [primary_path, bak, standby_path] {
-        let _ = std::fs::remove_file(p);
-    }
-}
-
-// Stage 18 slice 3: a replication slot holds WAL-ring truncation at a
-// standby's received LSN, survives a restart, and is invalidated once it
-// lags past `max_slot_retain_bytes`.
-#[test]
-fn replication_slots_hold_truncation_persist_and_reap() {
-    let path = unique_temp_path("repl-slots");
-    let engine = new_engine(&path);
-    engine
-        .execute("CREATE TABLE t (id INT NOT NULL PRIMARY KEY, v INT)")
-        .expect("create");
-    for i in 1..=10 {
-        engine
-            .execute(&format!("INSERT INTO t VALUES ({i}, {i})"))
-            .expect("insert");
-    }
-    let l1 = engine.storage().wal_flushed_lsn();
-    assert!(l1 > 0, "there is log to pin");
-
-    // A slot pinned at l1 holds the WAL head there even as the tail advances.
-    engine
-        .storage()
-        .try_register_repl_slot(7, l1)
-        .expect("register");
-    for i in 11..=30 {
-        engine
-            .execute(&format!("INSERT INTO t VALUES ({i}, {i})"))
-            .expect("insert");
-    }
-    engine.checkpoint().expect("checkpoint");
-    assert_eq!(
-        engine.storage().wal_head(),
-        l1,
-        "the slot pins the truncation floor at its LSN"
-    );
-
-    // Advancing the slot lets the floor follow.
-    let l2 = engine.storage().wal_flushed_lsn();
-    engine.storage().advance_repl_slot(7, l2);
-    engine.checkpoint().expect("checkpoint");
-    assert_eq!(
-        engine.storage().wal_head(),
-        l2,
-        "advancing the slot advances the floor"
-    );
-
-    // The slot's hold survives a restart (re-seeded from the superblock).
-    drop(engine);
-    let engine = Engine::new(Storage::open(path.clone()).expect("open")).expect("engine");
-    assert_eq!(
-        engine.storage().repl_slot_lsn(7),
-        Some(l2),
-        "the slot survives the restart"
-    );
-
-    // An explicit drop removes a slot (a standby that deregisters).
-    engine
-        .storage()
-        .try_register_repl_slot(8, l2)
-        .expect("register");
-    engine.storage().drop_repl_slot(8);
-    assert_eq!(
-        engine.storage().repl_slot_lsn(8),
-        None,
-        "an explicitly dropped slot is gone"
-    );
-
-    // Lagging past max_slot_retain invalidates the slot at the next checkpoint.
-    engine
-        .storage()
-        .set_max_slot_retain_bytes(64)
-        .expect("set cap");
-    for i in 31..=60 {
-        engine
-            .execute(&format!("INSERT INTO t VALUES ({i}, {i})"))
-            .expect("insert");
-    }
-    engine.checkpoint().expect("checkpoint");
-    assert_eq!(
-        engine.storage().repl_slot_lsn(7),
-        None,
-        "an over-lagging slot is invalidated"
-    );
-    assert!(
-        engine.storage().wal_head() > l2,
-        "the reclaimed floor advances past the dropped slot"
-    );
-
-    drop(engine);
-    let _ = std::fs::remove_file(path);
-}
-
-// A replication sender wakes on the group-commit flushed watch instead of
-// polling: a committed write advances the watch past the subscriber's last
-// seen value, and the durable watermark it re-reads covers the commit.
-#[test]
-fn the_flushed_watch_wakes_on_a_committed_write() {
-    let path = unique_temp_path("repl-watch");
-    let engine = new_engine(&path);
-    let mut rx = engine.storage().subscribe_wal_flushed();
-    let before = *rx.borrow_and_update();
-    engine
-        .execute("CREATE TABLE t (id INT NOT NULL PRIMARY KEY)")
-        .expect("create");
-    engine.execute("INSERT INTO t VALUES (1)").expect("insert");
-    assert!(
-        rx.has_changed().expect("sender alive"),
-        "a durable commit signals the watch"
-    );
-    let hint = *rx.borrow_and_update();
-    assert!(hint > before, "the watch value advances");
-    assert!(
-        engine.storage().wal_flushed_lsn() >= hint,
-        "the re-read watermark covers the hint"
-    );
-    drop(engine);
-    let _ = std::fs::remove_file(path);
-}
-
-// Slot registration is atomically fenced against truncation: a slot below
-// the WAL head is refused (the log it needs is gone — reseed), and the
-// table is bounded so a slot never silently fails to persist.
-#[test]
-fn slot_registration_rejects_below_head_and_a_full_table() {
-    let path = unique_temp_path("repl-slot-guards");
-    let engine = new_engine(&path);
-    engine
-        .execute("CREATE TABLE t (id INT NOT NULL PRIMARY KEY)")
-        .expect("create");
-    for i in 1..=10 {
-        engine
-            .execute(&format!("INSERT INTO t VALUES ({i})"))
-            .expect("insert");
-    }
-    engine.checkpoint().expect("checkpoint");
-    let head = engine.storage().wal_head();
-    assert!(head > 0, "the checkpoint truncated some log");
-    let err = engine
-        .storage()
-        .try_register_repl_slot(1, head - 1)
-        .expect_err("a below-head slot is refused");
-    assert!(
-        err.to_string().contains("reseed"),
-        "the error tells the operator to reseed: {err}"
-    );
-
-    let lsn = engine.storage().wal_flushed_lsn();
-    for id in 1..=8 {
-        engine
-            .storage()
-            .try_register_repl_slot(id, lsn)
-            .expect("register");
-    }
-    engine
-        .storage()
-        .try_register_repl_slot(9, lsn)
-        .expect_err("a ninth slot exceeds the persisted table");
-    // Re-registering an existing id is a reset, not a new slot.
-    engine
-        .storage()
-        .try_register_repl_slot(8, lsn)
-        .expect("re-register");
-    // A mid-entry LSN off the wire is refused: it would become the
-    // checkpoint truncation floor and, at the next restart, a WAL head the
-    // scan cannot decode (entries are >= 48 bytes and 8-aligned, so
-    // `fresh - 1` is never a boundary). Fresh writes keep it above the
-    // head so the boundary check — not the head fence — is what fires.
-    for i in 11..=13 {
-        engine
-            .execute(&format!("INSERT INTO t VALUES ({i})"))
-            .expect("insert");
-    }
-    let fresh = engine.storage().wal_flushed_lsn();
-    let err = engine
-        .storage()
-        .try_register_repl_slot(8, fresh - 1)
-        .expect_err("a mid-entry slot LSN is refused");
-    assert!(
-        err.to_string().contains("boundary"),
-        "the error names the misalignment: {err}"
-    );
-    // An ack for a never-registered (or reaped) slot must not create one.
-    engine.storage().advance_repl_slot(42, lsn);
-    assert_eq!(engine.storage().repl_slot_lsn(42), None);
-    drop(engine);
-    let _ = std::fs::remove_file(path);
-}
-
-// The replication epoch persists across a checkpoint (whose superblock
-// rebuild zeroes the reserved area — the checkpoint-wipe carry) and a
-// restart, and the retention-cap setter refuses a cap the ring cannot
-// honor.
-#[test]
-fn the_replication_epoch_persists_and_the_retain_cap_is_bounded() {
-    let path = unique_temp_path("repl-epoch");
-    let engine = new_engine(&path);
-    assert_eq!(engine.storage().epoch(), 0, "a fresh file is epoch 0");
-    engine.storage().set_epoch(5).expect("set epoch");
-    engine
-        .execute("CREATE TABLE t (id INT NOT NULL PRIMARY KEY)")
-        .expect("create");
-    engine.checkpoint().expect("checkpoint");
-    drop(engine);
-    let engine = Engine::new(Storage::open(path.clone()).expect("open")).expect("engine");
-    assert_eq!(
-        engine.storage().epoch(),
-        5,
-        "the epoch survives the checkpoint and the restart"
-    );
-    engine
-        .storage()
-        .set_max_slot_retain_bytes(u64::MAX - 1)
-        .expect_err("a cap at or above the usable ring is refused");
-    drop(engine);
-    let _ = std::fs::remove_file(path);
-}
-
-#[test]
-fn point_in_time_restore_stops_at_a_commit_timestamp() {
-    use crate::storage_layout::WAL_ENTRY_TYPE_REL;
-    use crate::wal::records::{REL_KIND_TXN_COMMIT, RelRecord};
-
-    let src = unique_temp_path("pitr-src");
-    let bak = unique_temp_path("pitr-bak");
-    let restored = unique_temp_path("pitr-restored");
-
-    let engine = new_engine(&src);
-    engine
-        .execute("CREATE TABLE t (id INT NOT NULL PRIMARY KEY)")
-        .expect("create");
-    // Three autocommitted inserts, spaced so their commit records land in
-    // distinct milliseconds; the sleeps make the timestamps separable.
-    engine.execute("INSERT INTO t VALUES (1)").expect("ins 1");
-    std::thread::sleep(std::time::Duration::from_millis(15));
-    engine.execute("INSERT INTO t VALUES (2)").expect("ins 2");
-    std::thread::sleep(std::time::Duration::from_millis(15));
-    engine.execute("INSERT INTO t VALUES (3)").expect("ins 3");
-    // The online full backup captures all three commits in its embedded log.
-    engine.storage().backup_full(&bak).expect("full backup");
-    drop(engine);
-
-    // Reopen the source so recovery's ring scan fills the replay cache, then
-    // read the commit timestamps in LSN order: create, insert-1, insert-2,
-    // insert-3. Stopping at insert-1's timestamp must keep row 1 (ts == stop
-    // is not "after") and undo rows 2 and 3 (ts > stop).
-    let storage = Storage::open(src.clone()).expect("reopen src");
-    let mut commit_ts: Vec<u64> = Vec::new();
-    for r in storage.replay_wal_entries().expect("replay") {
-        if r.entry_type != WAL_ENTRY_TYPE_REL {
-            continue;
-        }
-        let rec = RelRecord::decode(&r.payload).expect("decode");
-        if let Some(ts) = rec.commit_timestamp_millis() {
-            commit_ts.push(ts);
-        }
-    }
-    drop(storage);
-    assert!(
-        commit_ts.len() >= 4,
-        "create + three inserts commit; got {commit_ts:?}"
-    );
-    let stop_at = commit_ts[1]; // insert-1's commit timestamp
-    assert!(
-        commit_ts[2] > stop_at,
-        "insert-2 must commit strictly after insert-1 for a clean stop point: {commit_ts:?}"
-    );
-
-    // Point-in-time restore of the full backup, stopping after insert-1.
-    Storage::restore_full_with_logs(&restored, &bak, &[], Some(stop_at)).expect("pitr restore");
-    let engine2 = Engine::new(Storage::open(restored.clone()).expect("open")).expect("engine");
-    assert_eq!(
-        sql_rows(&engine2, "SELECT id FROM t ORDER BY id").1,
-        vec![vec![Some("1".into())]],
-        "only the commit at-or-before the stop point survives"
-    );
-    drop(engine2);
-
-    // The undo persists: a plain reopen (no stop point) still shows only
-    // row 1 — the losers were rolled back durably via CLRs, not re-derived
-    // from stop_at.
-    let engine3 = Engine::new(Storage::open(restored.clone()).expect("reopen")).expect("engine");
-    assert_eq!(
-        sql_rows(&engine3, "SELECT id FROM t ORDER BY id").1,
-        vec![vec![Some("1".into())]],
-        "point-in-time state survives a normal restart"
-    );
-    drop(engine3);
-
-    for p in [src, bak, restored] {
-        let _ = std::fs::remove_file(p);
-    }
-}
-
-#[test]
-fn restore_inspect_verbs_read_a_backup_without_restoring() {
-    let src = unique_temp_path("restinspect-src");
-    let bak = unique_temp_path("restinspect-bak");
-
-    let engine = new_engine(&src);
-    engine
-        .execute("CREATE TABLE t (id INT NOT NULL PRIMARY KEY, v INT)")
-        .expect("create");
-    for i in 1..=10 {
-        engine
-            .execute(&format!("INSERT INTO t VALUES ({i}, {i})"))
-            .expect("insert");
-    }
-    let baklit = bak.to_str().unwrap().replace('\'', "''");
-    assert!(
-        sql(
-            &engine,
-            &format!("BACKUP DATABASE truthdb TO DISK = '{baklit}'")
-        )["error"]
-            .is_null()
-    );
-
-    // HEADERONLY: exactly one metadata row; a full backup is BackupType 1.
-    let (cols, rows) = sql_rows(
-        &engine,
-        &format!("RESTORE HEADERONLY FROM DISK = '{baklit}'"),
-    );
-    assert_eq!(rows.len(), 1, "one header row");
-    assert!(cols.contains(&"BackupType".to_string()));
-    assert!(cols.contains(&"Checksum".to_string()));
-    let col = |name: &str| rows[0][cols.iter().position(|c| c == name).unwrap()].clone();
-    assert_eq!(col("BackupType"), Some("1".to_string()), "full backup");
-    assert_eq!(col("FormatVersion"), Some("2".to_string()));
-    assert_eq!(col("PageSize"), Some("4096".to_string()));
-
-    // FILELISTONLY: a data row ('D') and a log row ('L').
-    let (fcols, frows) = sql_rows(
-        &engine,
-        &format!("RESTORE FILELISTONLY FROM DISK = '{baklit}'"),
-    );
-    assert_eq!(fcols, vec!["LogicalName", "Type", "Size"]);
-    let types: Vec<Option<String>> = frows.iter().map(|r| r[1].clone()).collect();
-    assert_eq!(types, vec![Some("D".to_string()), Some("L".to_string())]);
-
-    // VERIFYONLY: a valid backup verifies with no error and opens no rowset.
-    let env = sql(
-        &engine,
-        &format!("RESTORE VERIFYONLY FROM DISK = '{baklit}'"),
-    );
-    assert!(env["error"].is_null(), "valid backup verifies: {env}");
-
-    // RESTORE DATABASE is offline-only: online it errors (3101).
-    assert_eq!(
-        sql_error_number(
-            &engine,
-            &format!("RESTORE DATABASE truthdb FROM DISK = '{baklit}'")
-        ),
-        3101
-    );
-
-    // Inside a transaction, restore is refused (3021), like BACKUP.
-    assert_eq!(
-        sql_error_number(
-            &engine,
-            &format!("BEGIN TRANSACTION; RESTORE VERIFYONLY FROM DISK = '{baklit}'")
-        ),
-        3021
-    );
-    drop(engine);
-
-    for p in [src, bak] {
-        let _ = std::fs::remove_file(p);
-    }
-}
-
-#[test]
-fn restore_verifyonly_rejects_a_corrupt_or_missing_backup() {
-    let src = unique_temp_path("restverify-src");
-    let bak = unique_temp_path("restverify-bak");
-    let engine = new_engine(&src);
-    engine
-        .execute("CREATE TABLE t (id INT NOT NULL PRIMARY KEY)")
-        .expect("create");
-    engine.execute("INSERT INTO t VALUES (1)").expect("insert");
-    let baklit = bak.to_str().unwrap().replace('\'', "''");
-    assert!(
-        sql(
-            &engine,
-            &format!("BACKUP DATABASE truthdb TO DISK = '{baklit}'")
-        )["error"]
-            .is_null()
-    );
-
-    let pristine = std::fs::read(&bak).expect("read bak");
-    let verify = |bytes: &[u8]| {
-        std::fs::write(&bak, bytes).expect("write");
-        sql_error_number(
-            &engine,
-            &format!("RESTORE VERIFYONLY FROM DISK = '{baklit}'"),
-        )
-    };
-
-    // Flip a payload byte mid-file: it no longer matches its xxh64, so
-    // VERIFYONLY reports the restore terminating abnormally (3013).
-    let mut payload_flip = pristine.clone();
-    let mid = payload_flip.len() / 2;
-    payload_flip[mid] ^= 0xFF;
-    assert_eq!(verify(&payload_flip), 3013, "a flipped payload byte");
-
-    // Corrupt the header block's length field (its high byte, outside the
-    // checksum): recovery must report 3013, not allocate ~u64::MAX and crash.
-    let mut len_flip = pristine.clone();
-    len_flip[19] = 0xFF;
-    assert_eq!(
-        verify(&len_flip),
-        3013,
-        "a corrupt block length is not a crash"
-    );
-
-    // A missing file errors cleanly, not a panic.
-    assert_eq!(
-        sql_error_number(
-            &engine,
-            "RESTORE VERIFYONLY FROM DISK = '/nonexistent/nope.bak'"
-        ),
-        3013
-    );
-    drop(engine);
-    for p in [src, bak] {
-        let _ = std::fs::remove_file(p);
-    }
-}
-
-#[test]
-fn backup_under_concurrent_write_load_restores_to_a_consistent_prefix() {
-    use std::sync::Arc;
-    let src = unique_temp_path("bul-src");
-    let bak = unique_temp_path("bul-bak");
-    let restored = unique_temp_path("bul-restored");
-
-    let engine = Arc::new(new_engine(&src));
-    engine
-        .execute("CREATE TABLE t (id INT NOT NULL PRIMARY KEY)")
-        .expect("create");
-    // A baseline committed BEFORE the backup starts: the restore must contain
-    // at least these 20 rows.
-    for i in 1..=20 {
-        engine
-            .execute(&format!("INSERT INTO t VALUES ({i})"))
-            .expect("baseline insert");
-    }
-
-    // A writer thread commits ids 21..=80 while the backup runs, so the fuzzy
-    // page copy interleaves with live commits (online backup under load). The
-    // count is kept modest so the test stays short and does not hold io_uring
-    // resources long enough to pressure other tests running in parallel.
-    let writer = {
-        let engine = Arc::clone(&engine);
-        std::thread::spawn(move || {
-            for i in 21..=80 {
-                engine
-                    .execute(&format!("INSERT INTO t VALUES ({i})"))
-                    .expect("concurrent insert");
-            }
-        })
-    };
-    engine
-        .storage()
-        .backup_full(&bak)
-        .expect("online backup under write load");
-    writer.join().expect("writer thread");
-    drop(engine);
-
-    // The restore recovers to a single consistent LSN (backup_end). Commits
-    // are serialized in id order, so the restored ids must be a CONTIGUOUS
-    // prefix 1..=k — a gap would mean a torn page, an id past k an
-    // uncommitted write leaked in. k lies between the 20 baseline rows and
-    // the writer's max (80).
-    Storage::restore_full(&restored, &bak).expect("restore");
-    let engine2 = Engine::new(Storage::open(restored.clone()).expect("open")).expect("engine");
-    let (_, rows) = sql_rows(&engine2, "SELECT id FROM t ORDER BY id");
-    let ids: Vec<i64> = rows
-        .iter()
-        .map(|r| r[0].as_ref().unwrap().parse().unwrap())
-        .collect();
-    assert!(
-        (20..=80).contains(&ids.len()),
-        "restored count is between the baseline and the writer's max, got {}",
-        ids.len()
-    );
-    for (i, id) in ids.iter().enumerate() {
-        assert_eq!(
-            *id,
-            i as i64 + 1,
-            "restored ids are a contiguous prefix (no gaps, no torn/phantom rows): {ids:?}"
-        );
-    }
-    drop(engine2);
-    for p in [src, bak, restored] {
-        let _ = std::fs::remove_file(p);
-    }
-}
-
-#[test]
-fn a_failed_backup_leaves_the_database_writable_and_backuppable() {
-    let src = unique_temp_path("killmid-src");
-    let good = unique_temp_path("killmid-good");
-    let restored = unique_temp_path("killmid-restored");
-    let engine = new_engine(&src);
-    engine
-        .execute("CREATE TABLE t (id INT NOT NULL PRIMARY KEY)")
-        .expect("create");
-    engine.execute("INSERT INTO t VALUES (1)").expect("insert");
-
-    // A backup to a nonexistent directory fails mid-flight: the hold is armed
-    // (begin_backup succeeded), then write_backup's File::create errors. The
-    // RAII hold guard must still release, or WAL truncation freezes and writes
-    // eventually wedge.
-    assert_eq!(
-        sql_error_number(
-            &engine,
-            "BACKUP DATABASE truthdb TO DISK = '/nonexistent-truthdb-dir/b.bak'"
-        ),
-        3013
-    );
-
-    // The database is unharmed: writes still work...
-    engine
-        .execute("INSERT INTO t VALUES (2)")
-        .expect("insert after a failed backup");
-    // ...and a fresh backup to a good path succeeds — which it could not if
-    // the failed backup's hold or single-flight guard were still set.
-    let goodlit = good.to_str().unwrap().replace('\'', "''");
-    assert!(
-        sql(
-            &engine,
-            &format!("BACKUP DATABASE truthdb TO DISK = '{goodlit}'")
-        )["error"]
-            .is_null()
-    );
-    drop(engine);
-
-    // And that good backup restores the surviving rows.
-    Storage::restore_full(&restored, &good).expect("restore");
-    let engine2 = Engine::new(Storage::open(restored.clone()).expect("open")).expect("engine");
-    assert_eq!(
-        sql_rows(&engine2, "SELECT id FROM t ORDER BY id").1,
-        vec![vec![Some("1".into())], vec![Some("2".into())]],
-        "the post-failure database backs up and restores intact"
-    );
-    drop(engine2);
-    for p in [src, good, restored] {
-        let _ = std::fs::remove_file(p);
-    }
-}
-
-#[test]
-fn recovery_redo_is_idempotent_across_repeated_reopens() {
-    // Redo is the resumable core of replication (a standby applies it as
-    // records arrive): re-running it must be a no-op, gated by each page's
-    // LSN. Reopening the same database repeatedly re-runs redo over the whole
-    // log each time; the committed state must survive unchanged.
-    let path = unique_temp_path("redo-idempotent");
-    {
-        let engine = new_engine(&path);
-        engine
-            .execute("CREATE TABLE t (id INT NOT NULL PRIMARY KEY, v INT)")
-            .expect("create");
-        for i in 1..=30 {
-            engine
-                .execute(&format!("INSERT INTO t VALUES ({i}, {i})"))
-                .expect("insert");
-        }
-        engine
-            .execute("UPDATE t SET v = v + 100 WHERE id <= 10")
-            .expect("update");
-    }
-    let expected = {
-        let engine = Engine::new(Storage::open(path.clone()).expect("open")).expect("engine");
-        sql_rows(&engine, "SELECT id, v FROM t ORDER BY id").1
-    };
-    // Reopen several more times: each reopen re-runs redo over the full log.
-    // The state is invariant — proving redo re-application is idempotent.
-    for _ in 0..3 {
-        let engine = Engine::new(Storage::open(path.clone()).expect("reopen")).expect("engine");
-        assert_eq!(
-            sql_rows(&engine, "SELECT id, v FROM t ORDER BY id").1,
-            expected,
-            "redo re-application over the whole log is a no-op"
-        );
-    }
-    let _ = std::fs::remove_file(path);
-}
-
-#[test]
-fn commit_records_carry_a_recent_wall_clock_timestamp() {
-    use crate::storage_layout::WAL_ENTRY_TYPE_REL;
-    use crate::wal::records::{REL_KIND_TXN_COMMIT, RelRecord};
-    let now_ms = || {
-        std::time::SystemTime::now()
-            .duration_since(std::time::UNIX_EPOCH)
-            .unwrap()
-            .as_millis() as u64
-    };
-    let path = unique_temp_path("commit-ts");
-    let before = now_ms();
-    {
-        let engine = new_engine(&path);
-        engine
-            .execute("CREATE TABLE t (id INT NOT NULL PRIMARY KEY)")
-            .expect("create");
-        engine.execute("INSERT INTO t VALUES (1)").expect("insert"); // autocommits
-    }
-    let after = now_ms();
-
-    // Reopen so recovery's ring scan populates the replay cache; some
-    // committed transaction's record carries a v2 entry with a wall-clock
-    // timestamp in [before, after] (for point-in-time restore).
-    let storage = Storage::open(path.clone()).expect("reopen");
-    let records = storage.replay_wal_entries().expect("replay");
-    let mut found = false;
-    for r in records {
-        if r.entry_type != WAL_ENTRY_TYPE_REL {
-            continue;
-        }
-        let rec = RelRecord::decode(&r.payload).expect("decode");
-        if rec.kind == REL_KIND_TXN_COMMIT && rec.redo.len() >= 8 {
-            assert_eq!(r.entry_version, 2, "new commit records are entry version 2");
-            let ts = u64::from_le_bytes(rec.redo[..8].try_into().unwrap());
-            if before <= ts && ts <= after {
-                found = true;
-            }
-        }
-    }
-    assert!(
-        found,
-        "a commit record carried a timestamp in [{before}, {after}]"
-    );
-    drop(storage);
-    let _ = std::fs::remove_file(path);
-}
-
-#[test]
-fn backup_log_requires_the_full_recovery_model() {
-    let path = unique_temp_path("backuplog-simple");
-    let engine = new_engine(&path);
-    // SIMPLE (the default): BACKUP LOG is 4208.
-    assert_eq!(
-        sql_error_number(
-            &engine,
-            "BACKUP LOG truthdb TO DISK = '/tmp/truthdb-never.trn'"
-        ),
-        4208
-    );
-    drop(engine);
-    let _ = std::fs::remove_file(path);
-}
-
-#[test]
-fn full_recovery_holds_the_log_until_backup_log() {
-    let path = unique_temp_path("backuplog-hold");
-    let trn = unique_temp_path("backuplog-hold-trn");
-    let engine = new_engine(&path);
-    engine
-        .execute("ALTER DATABASE CURRENT SET RECOVERY FULL")
-        .expect("full");
-    let marker = engine.storage().last_log_backup_lsn();
-    engine
-        .execute("CREATE TABLE t (id INT NOT NULL PRIMARY KEY, v INT)")
-        .expect("create");
-    for i in 1..=50 {
-        engine
-            .execute(&format!("INSERT INTO t VALUES ({i}, {i})"))
-            .expect("insert");
-    }
-    // A checkpoint cannot truncate past the log-backup floor under FULL.
-    engine
-        .storage()
-        .write_checkpoint(b"cp", 1, 2, 1)
-        .expect("checkpoint");
-    assert!(
-        engine.storage().wal_head() <= marker,
-        "FULL pins the head at the log-backup floor"
-    );
-
-    // BACKUP LOG advances the floor, so a later checkpoint reclaims the log.
-    let lit = trn.to_str().unwrap().replace('\'', "''");
-    assert!(sql(&engine, &format!("BACKUP LOG truthdb TO DISK = '{lit}'"))["error"].is_null());
-    assert!(engine.storage().last_log_backup_lsn() > marker);
-    engine
-        .storage()
-        .write_checkpoint(b"cp2", 2, 3, 2)
-        .expect("checkpoint");
-    assert!(
-        engine.storage().wal_head() > marker,
-        "after BACKUP LOG the checkpoint reclaims past the old floor"
-    );
-    drop(engine);
-    for p in [path, trn] {
-        let _ = std::fs::remove_file(p);
-    }
-}
-
-#[test]
-fn log_backup_marker_and_hold_survive_reopen() {
-    let path = unique_temp_path("backuplog-reopen");
-    let trn = unique_temp_path("backuplog-reopen-trn");
-    let engine = new_engine(&path);
-    engine
-        .execute("ALTER DATABASE CURRENT SET RECOVERY FULL")
-        .expect("full");
-    engine
-        .execute("CREATE TABLE t (id INT NOT NULL PRIMARY KEY)")
-        .expect("create");
-    engine.execute("INSERT INTO t VALUES (1)").expect("insert");
-    let lit = trn.to_str().unwrap().replace('\'', "''");
-    assert!(sql(&engine, &format!("BACKUP LOG truthdb TO DISK = '{lit}'"))["error"].is_null());
-    let marker = engine.storage().last_log_backup_lsn();
-    drop(engine);
-
-    let engine2 = Engine::new(Storage::open(path.clone()).expect("reopen")).expect("engine");
-    assert_eq!(
-        engine2.storage().last_log_backup_lsn(),
-        marker,
-        "the marker persisted"
-    );
-    assert_eq!(
-        engine2.storage().log_backup_hold(),
-        Some(marker),
-        "the hold re-registered on open"
-    );
-    assert_eq!(
-        sql_rows(&engine2, "SELECT recovery_model_desc FROM sys.databases").1,
-        vec![vec![Some("FULL".into())]]
-    );
-    drop(engine2);
-    for p in [path, trn] {
-        let _ = std::fs::remove_file(p);
-    }
-}
-
-#[test]
-fn full_recovery_ring_full_reports_9002() {
-    let path = unique_temp_path("backuplog-9002");
-    // A small ring so the un-backed-up log fills it quickly.
-    let storage = Storage::create_with_wal_bounds(
-        path.clone(),
-        test_storage_options(),
-        128 * 1024,
-        128 * 1024,
-    )
-    .expect("create");
-    let engine = Engine::new(storage).expect("engine");
-    engine
-        .execute("ALTER DATABASE CURRENT SET RECOVERY FULL")
-        .expect("full");
-    engine
-        .execute("CREATE TABLE t (id INT NOT NULL PRIMARY KEY, v NVARCHAR(120))")
-        .expect("create");
-    // FULL pins the head at the enable point, so with no BACKUP LOG the ring
-    // fills and a write eventually reports 9002 (log full).
-    let mut hit_9002 = false;
-    for i in 0..4000 {
-        let env = sql(
-            &engine,
-            &format!("INSERT INTO t VALUES ({i}, '{}')", "x".repeat(100)),
-        );
-        if let Some(n) = env["error"]["number"].as_i64() {
-            assert_eq!(n, 9002, "ring-full under FULL reports 9002, got {env}");
-            hit_9002 = true;
-            break;
-        }
-    }
-    assert!(hit_9002, "the ring filled and reported 9002");
-    drop(engine);
     let _ = std::fs::remove_file(path);
 }
 
@@ -5131,6 +2317,12 @@ fn sql_batch_keeps_earlier_results_before_an_error() {
 
 use crate::engine::{BatchOutcome, StatementResult, TxnContext};
 use crate::relstore::types::Datum;
+
+/// Runs a SQL batch through the session path with a persistent transaction
+/// context (as a TDS connection would), returning the typed outcome.
+fn batch(engine: &Engine, ctx: &mut TxnContext, sql: &str) -> BatchOutcome {
+    engine.sql_batch(sql, ctx).expect("sql_batch")
+}
 
 /// The integer `id` column (column 0) of the first rowset in an outcome.
 fn ids(outcome: &BatchOutcome) -> Vec<i32> {
